Summarise Matrix Population Models

Calculates and summarises various metrics from matrix population models (MPMs) including dimension (= age in Leslie matrices), lambda values, maximum fecundity values, maximum growth/survival transition probabilities, and minimum non-zero growth/survival transition probabilities

Usage

summarise_mpms(x)

Arguments

x

A compadreDB object containing matrix population models, typically in a format compatible with matA, matU, and matF functions.

Value

This function prints summaries of the following metrics:

lambda values: The lambda values (dominant eigenvalues) of the A matrices.

max F values: The maximum values from the F matrices.

max U values: The maximum values from the U matrices.

minimum non-zero U values: The minimum non-zero values from the U matrices.

See also

Examples

mats <- rand_lefko_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, output = "Type1"
)

summarise_mpms(mats)
#> Summary of matrix dimension:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>       5       5       5       5       5       5 
#> Summary of lambda values:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   1.000   1.060   1.173   1.194   1.232   1.628 
#> 
#> Summary of maximum F values:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>      10      10      10      10      10      10 
#> 
#> Summary of maximum U values:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.9194  0.9372  0.9687  0.9590  0.9761  0.9869 
#> 
#> Summary of minimum non-zero U values:
#>      Min.   1st Qu.    Median      Mean   3rd Qu.      Max. 
#> 0.0006678 0.0074466 0.0190447 0.0341075 0.0438259 0.1366390