Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable lambda values. The function attempts to find matrices that fulfill the lambda criteria, discarding unacceptable matrices. If it takes more than 1000 attempts to find a suitable matrix model, then an error is produced.

Usage

generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99
)

Arguments

n

The number of MPMs to generate. Default is 10.

lower_lambda

The lower bound for acceptable lambda values. Default is 0.9.

upper_lambda

The upper bound for acceptable lambda values. Default is 1.1.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,] 0.46959019 0.0000000 2.00000000 9.00000000 13.0000000
#> [2,] 0.06208989 0.5353928 0.00000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.1655550 0.02415932 0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.74073556 0.76087838  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.08745141  0.8588216
#> 
#> $A_list[[2]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1637908 0.0000000 6.0000000 9.0000000 11.0000000
#> [2,] 0.1820130 0.3266282 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1201711 0.2278926 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3388107 0.0551089  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6111687  0.7516722
#> 
#> $A_list[[3]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.21493275 0.0000000 4.00000000 6.0000000 8.0000000
#> [2,] 0.03999446 0.2906202 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1958827 0.70105409 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.06471572 0.4448149 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.4108997 0.9636414
#> 
#> $A_list[[4]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.04767234 0.00000000 4.0000000 10.0000000 8.0000000
#> [2,] 0.46547580 0.65443741 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.03106457 0.5236817  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2490691  0.1855063 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.7165077 0.9348314
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.15027486 0.00000000 2.0000000 11.0000000 9.0000000
#> [2,] 0.09914183 0.27779157 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.00966585 0.3919230  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2236831  0.8317754 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.1314628 0.9736579
#> 
#> $A_list[[6]]
#>            [,1]      [,2]       [,3]        [,4]      [,5]
#> [1,] 0.24966518 0.0000000 3.00000000 11.00000000 9.0000000
#> [2,] 0.01067982 0.1639693 0.00000000  0.00000000 0.0000000
#> [3,] 0.00000000 0.6896182 0.01409515  0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.94692777  0.91181810 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000  0.05884282 0.9714646
#> 
#> $A_list[[7]]
#>             [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.197324394 0.0000000 3.0000000 10.0000000 10.0000000
#> [2,] 0.006074399 0.1033257 0.0000000  0.0000000  0.0000000
#> [3,] 0.000000000 0.2749083 0.4577717  0.0000000  0.0000000
#> [4,] 0.000000000 0.0000000 0.1307701  0.3286445  0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000  0.5040002  0.9461408
#> 
#> $A_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.49046382 0.00000000 2.0000000 7.0000000 8.0000000
#> [2,] 0.03707287 0.05319779 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.62634254 0.3704960 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3517029 0.2565015 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6099138 0.9488008
#> 
#> $A_list[[9]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1147491 0.00000000 3.0000000 3.0000000 9.0000000
#> [2,] 0.2679386 0.52503660 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.01317803 0.3052731 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.3123143 0.2300555 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.6822216 0.9895889
#> 
#> $A_list[[10]]
#>           [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.3618345 0.0000000 1.00000000 8.00000000 8.0000000
#> [2,] 0.1250533 0.2773141 0.00000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.3120913 0.51516988 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.07960787 0.69899711 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.09891104 0.8910903
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.46959019 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.06208989 0.5353928 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1655550 0.02415932 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.74073556 0.76087838 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.08745141 0.8588216
#> 
#> $U_list[[2]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1637908 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1820130 0.3266282 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1201711 0.2278926 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3388107 0.0551089 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6111687 0.7516722
#> 
#> $U_list[[3]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.21493275 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.03999446 0.2906202 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1958827 0.70105409 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.06471572 0.4448149 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.4108997 0.9636414
#> 
#> $U_list[[4]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.04767234 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.46547580 0.65443741 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03106457 0.5236817 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2490691 0.1855063 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.7165077 0.9348314
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.15027486 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.09914183 0.27779157 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.00966585 0.3919230 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2236831 0.8317754 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1314628 0.9736579
#> 
#> $U_list[[6]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.24966518 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.01067982 0.1639693 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.6896182 0.01409515 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.94692777 0.91181810 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.05884282 0.9714646
#> 
#> $U_list[[7]]
#>             [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.197324394 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.006074399 0.1033257 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.2749083 0.4577717 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.1307701 0.3286445 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.5040002 0.9461408
#> 
#> $U_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.49046382 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03707287 0.05319779 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.62634254 0.3704960 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3517029 0.2565015 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6099138 0.9488008
#> 
#> $U_list[[9]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1147491 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2679386 0.52503660 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.01317803 0.3052731 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.3123143 0.2300555 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.6822216 0.9895889
#> 
#> $U_list[[10]]
#>           [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.3618345 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.1250533 0.2773141 0.00000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.3120913 0.51516988 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.07960787 0.69899711 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.09891104 0.8910903
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    9   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    9   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    6    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   10    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   11    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   11    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    3    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>