Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfill the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

an integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE
)
#> Error in generate_mpm_set(n = 10, n_stages = 5, fecundity = c(0, 0, 4,     8, 10), archetype = 4, split = FALSE): If split is FALSE, then by_type must also be FALSE

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(
  fun = "lambda", arg = NA, lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.59172410 0.00000000 3.0000000 7.0000000 13.0000000
#> [2,] 0.03928654 0.64623186 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.02707035 0.2940232 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.5196510 0.6367752  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2315941  0.9254113
#> 
#> $A_list[[2]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.24626067 0.0000000 5.00000000 8.0000000 12.0000000
#> [2,] 0.07629902 0.2497351 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1290864 0.66495243 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.09219072 0.2893268  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.4996649  0.8650195
#> 
#> $A_list[[3]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.20478944 0.0000000 4.00000000 8.0000000 12.0000000
#> [2,] 0.01509631 0.3194143 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3397877 0.08749785 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.61277615 0.2539657  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5771964  0.9680015
#> 
#> $A_list[[4]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.49141268 0.0000000 3.0000000 10.0000000 11.0000000
#> [2,] 0.02319536 0.3173920 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.3038625 0.6017887  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1626274  0.1808382  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.6091564  0.9047155
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.07627263 0.00000000 6.0000000 13.0000000 15.0000000
#> [2,] 0.06781778 0.26903098 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.06399574 0.5515509  0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.1701671  0.6095316  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.1790399  0.8856996
#> 
#> $A_list[[6]]
#>           [,1]        [,2]       [,3]       [,4]       [,5]
#> [1,] 0.1860569 0.000000000 4.00000000 11.0000000 15.0000000
#> [2,] 0.2342026 0.448128178 0.00000000  0.0000000  0.0000000
#> [3,] 0.0000000 0.008416447 0.08124431  0.0000000  0.0000000
#> [4,] 0.0000000 0.000000000 0.67295693  0.6248886  0.0000000
#> [5,] 0.0000000 0.000000000 0.00000000  0.3212519  0.9841611
#> 
#> $A_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.10871782 0.0000000 2.0000000 6.0000000 13.000000
#> [2,] 0.03368915 0.1020122 0.0000000 0.0000000  0.000000
#> [3,] 0.00000000 0.1311565 0.2535163 0.0000000  0.000000
#> [4,] 0.00000000 0.0000000 0.1828480 0.2229901  0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4900117  0.953543
#> 
#> $A_list[[8]]
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.03466013 0.00000000 4.00000000 6.0000000 14.0000000
#> [2,] 0.10623234 0.04276315 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.29450065 0.69949682 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.00435283 0.1876286  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6661410  0.9096006
#> 
#> $A_list[[9]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08147748 0.0000000 6.0000000 9.0000000 7.0000000
#> [2,] 0.05177217 0.4443598 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1201678 0.6047314 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1842868 0.3216026 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5020081 0.9206748
#> 
#> $A_list[[10]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.5055821 0.0000000 2.00000000 7.0000000 14.0000000
#> [2,] 0.1557725 0.6568405 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.0128138 0.71495796 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.01988592 0.1540314  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.8262843  0.9833724
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.59172410 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03928654 0.64623186 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02707035 0.2940232 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.5196510 0.6367752 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2315941 0.9254113
#> 
#> $U_list[[2]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.24626067 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.07629902 0.2497351 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1290864 0.66495243 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.09219072 0.2893268 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.4996649 0.8650195
#> 
#> $U_list[[3]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.20478944 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.01509631 0.3194143 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3397877 0.08749785 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.61277615 0.2539657 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5771964 0.9680015
#> 
#> $U_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.49141268 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02319536 0.3173920 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3038625 0.6017887 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1626274 0.1808382 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6091564 0.9047155
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07627263 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.06781778 0.26903098 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.06399574 0.5515509 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1701671 0.6095316 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1790399 0.8856996
#> 
#> $U_list[[6]]
#>           [,1]        [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1860569 0.000000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.2342026 0.448128178 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.008416447 0.08124431 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.67295693 0.6248886 0.0000000
#> [5,] 0.0000000 0.000000000 0.00000000 0.3212519 0.9841611
#> 
#> $U_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]     [,5]
#> [1,] 0.10871782 0.0000000 0.0000000 0.0000000 0.000000
#> [2,] 0.03368915 0.1020122 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.1311565 0.2535163 0.0000000 0.000000
#> [4,] 0.00000000 0.0000000 0.1828480 0.2229901 0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4900117 0.953543
#> 
#> $U_list[[8]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.03466013 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.10623234 0.04276315 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.29450065 0.69949682 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.00435283 0.1876286 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6661410 0.9096006
#> 
#> $U_list[[9]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08147748 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.05177217 0.4443598 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1201678 0.6047314 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1842868 0.3216026 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5020081 0.9206748
#> 
#> $U_list[[10]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.5055821 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1557725 0.6568405 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.0128138 0.71495796 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.01988592 0.1540314 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.8262843 0.9833724
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    7   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   13   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   11   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    6   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    9    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(
  fun = "eigs", arg = "lambda", lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>             [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.007874915 0.0000000 2.00000000 3.0000000 16.0000000
#> [2,] 0.007650961 0.1749329 0.00000000 0.0000000  0.0000000
#> [3,] 0.000000000 0.3340613 0.01670727 0.0000000  0.0000000
#> [4,] 0.000000000 0.0000000 0.87107796 0.5002390  0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.4105957  0.9366168
#> 
#> $A_list[[2]]
#>            [,1]        [,2]       [,3]      [,4]      [,5]
#> [1,] 0.18194573 0.000000000 2.00000000 6.0000000 6.0000000
#> [2,] 0.01738281 0.509414004 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.001869208 0.58665037 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.02794871 0.2439485 0.0000000
#> [5,] 0.00000000 0.000000000 0.00000000 0.6009070 0.9336844
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.18766701 0.00000000 4.0000000 4.00000000 10.0000000
#> [2,] 0.06583538 0.35863349 0.0000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.09893181 0.3882112 0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.2392777 0.90508120  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.07384031  0.9833391
#> 
#> $A_list[[4]]
#>            [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,] 0.04450171 0.0000000 2.00000000 12.0000000 11.0000000
#> [2,] 0.02967342 0.1403584 0.00000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.1518966 0.03907807  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.43409523  0.1589040  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000  0.6538104  0.8765643
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]        [,4]       [,5]
#> [1,] 0.42759639 0.00000000 6.0000000 11.00000000 10.0000000
#> [2,] 0.03622153 0.04718323 0.0000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.45548549 0.1821243  0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.3650009  0.87153157  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.04666619  0.9370567
#> 
#> $A_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.19690390 0.0000000 4.0000000 5.0000000 9.0000000
#> [2,] 0.09884488 0.1757495 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3778564 0.3290832 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2322539 0.6126456 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3144943 0.9289182
#> 
#> $A_list[[7]]
#>            [,1]      [,2]        [,3]        [,4]       [,5]
#> [1,] 0.07961651 0.0000000 3.000000000 14.00000000 12.0000000
#> [2,] 0.03913764 0.6197591 0.000000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.1255677 0.004200616  0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.820923863  0.83134063  0.0000000
#> [5,] 0.00000000 0.0000000 0.000000000  0.05401238  0.9305837
#> 
#> $A_list[[8]]
#>           [,1]      [,2]       [,3]        [,4]      [,5]
#> [1,] 0.2206521 0.0000000 3.00000000 14.00000000 7.0000000
#> [2,] 0.2052189 0.2987133 0.00000000  0.00000000 0.0000000
#> [3,] 0.0000000 0.1809668 0.52536120  0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.06262531  0.02975681 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000  0.57502242 0.6554696
#> 
#> $A_list[[9]]
#>           [,1]      [,2]        [,3]       [,4]      [,5]
#> [1,] 0.1152385 0.0000000 5.000000000 8.00000000 8.0000000
#> [2,] 0.2749706 0.2961082 0.000000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.1207265 0.702552288 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.005673965 0.03578905 0.0000000
#> [5,] 0.0000000 0.0000000 0.000000000 0.90489707 0.9453126
#> 
#> $A_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.33185284 0.0000000 5.0000000 9.0000000 13.0000000
#> [2,] 0.01607329 0.5320239 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1973191 0.1952510 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.7169374 0.1025489  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.8583364  0.9865513
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>             [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.007874915 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.007650961 0.1749329 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.3340613 0.01670727 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.87107796 0.5002390 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.4105957 0.9366168
#> 
#> $U_list[[2]]
#>            [,1]        [,2]       [,3]      [,4]      [,5]
#> [1,] 0.18194573 0.000000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.01738281 0.509414004 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.001869208 0.58665037 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.02794871 0.2439485 0.0000000
#> [5,] 0.00000000 0.000000000 0.00000000 0.6009070 0.9336844
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.18766701 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.06583538 0.35863349 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.09893181 0.3882112 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2392777 0.90508120 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.07384031 0.9833391
#> 
#> $U_list[[4]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04450171 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.02967342 0.1403584 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1518966 0.03907807 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.43409523 0.1589040 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.6538104 0.8765643
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.42759639 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.03622153 0.04718323 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.45548549 0.1821243 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3650009 0.87153157 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.04666619 0.9370567
#> 
#> $U_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.19690390 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.09884488 0.1757495 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3778564 0.3290832 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2322539 0.6126456 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3144943 0.9289182
#> 
#> $U_list[[7]]
#>            [,1]      [,2]        [,3]       [,4]      [,5]
#> [1,] 0.07961651 0.0000000 0.000000000 0.00000000 0.0000000
#> [2,] 0.03913764 0.6197591 0.000000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1255677 0.004200616 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.820923863 0.83134063 0.0000000
#> [5,] 0.00000000 0.0000000 0.000000000 0.05401238 0.9305837
#> 
#> $U_list[[8]]
#>           [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.2206521 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.2052189 0.2987133 0.00000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.1809668 0.52536120 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.06262531 0.02975681 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.57502242 0.6554696
#> 
#> $U_list[[9]]
#>           [,1]      [,2]        [,3]       [,4]      [,5]
#> [1,] 0.1152385 0.0000000 0.000000000 0.00000000 0.0000000
#> [2,] 0.2749706 0.2961082 0.000000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.1207265 0.702552288 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.005673965 0.03578905 0.0000000
#> [5,] 0.0000000 0.0000000 0.000000000 0.90489707 0.9453126
#> 
#> $U_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.33185284 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01607329 0.5320239 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1973191 0.1952510 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.7169374 0.1025489 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.8583364 0.9865513
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    3   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    4   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   12   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   11   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    5    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   14   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   14    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    9   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2132965 0.0000000 4.00000000 6.0000000 12.0000000
#> [2,] 0.1778327 0.5329060 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1099794 0.04235969 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.71970067 0.7168085  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.1609096  0.9600961
#> 
#> $A_list[[2]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.62190475 0.00000000 2.0000000 6.0000000 5.0000000
#> [2,] 0.08912391 0.05969362 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.69193367 0.6158058 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1541629 0.3566483 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5375225 0.9620333
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05674409 0.00000000 3.0000000 5.0000000 6.0000000
#> [2,] 0.12531586 0.28327981 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03299347 0.1884972 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2776971 0.1772986 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.4830728 0.7448259
#> 
#> $A_list[[4]]
#>            [,1]       [,2]      [,3]        [,4]       [,5]
#> [1,] 0.44665734 0.00000000 4.0000000 7.000000000 13.0000000
#> [2,] 0.05580094 0.05544291 0.0000000 0.000000000  0.0000000
#> [3,] 0.00000000 0.58806936 0.2956930 0.000000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.3857997 0.002040718  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.773762602  0.8690791
#> 
#> $A_list[[5]]
#>            [,1]      [,2]        [,3]      [,4]       [,5]
#> [1,] 0.24211557 0.0000000 5.000000000 6.0000000 11.0000000
#> [2,] 0.08827778 0.1851573 0.000000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1919922 0.407831020 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.007134864 0.4682320  0.0000000
#> [5,] 0.00000000 0.0000000 0.000000000 0.0560379  0.9668042
#> 
#> $A_list[[6]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05373296 0.00000000 1.0000000 9.0000000 7.0000000
#> [2,] 0.02451688 0.10170820 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.05681361 0.2203045 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1642305 0.2003433 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2685748 0.9532752
#> 
#> $A_list[[7]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.0608077 0.0000000 5.0000000 9.0000000 13.0000000
#> [2,] 0.4933830 0.4604336 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.2315650 0.3848140 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.4667048 0.1975581  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6653989  0.8872771
#> 
#> $A_list[[8]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.47685793 0.0000000 3.0000000 8.00000000 9.0000000
#> [2,] 0.01248983 0.3794055 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2989584 0.4294066 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3782046 0.05694344 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.82013455 0.9702802
#> 
#> $A_list[[9]]
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1240807 0.0000000 4.0000000 4.00000000 16.0000000
#> [2,] 0.2229549 0.6060821 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.0986676 0.6695799 0.00000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.2749632 0.86313873  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.09419495  0.9897981
#> 
#> $A_list[[10]]
#>           [,1]        [,2]      [,3]      [,4]       [,5]
#> [1,] 0.3255609 0.000000000 3.0000000 9.0000000 14.0000000
#> [2,] 0.3290727 0.721012443 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.006760425 0.2381221 0.0000000  0.0000000
#> [4,] 0.0000000 0.000000000 0.5494698 0.4816385  0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.3832743  0.9188889
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2132965 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1778327 0.5329060 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1099794 0.04235969 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.71970067 0.7168085 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.1609096 0.9600961
#> 
#> $U_list[[2]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.62190475 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.08912391 0.05969362 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.69193367 0.6158058 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1541629 0.3566483 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5375225 0.9620333
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05674409 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.12531586 0.28327981 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03299347 0.1884972 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2776971 0.1772986 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.4830728 0.7448259
#> 
#> $U_list[[4]]
#>            [,1]       [,2]      [,3]        [,4]      [,5]
#> [1,] 0.44665734 0.00000000 0.0000000 0.000000000 0.0000000
#> [2,] 0.05580094 0.05544291 0.0000000 0.000000000 0.0000000
#> [3,] 0.00000000 0.58806936 0.2956930 0.000000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3857997 0.002040718 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.773762602 0.8690791
#> 
#> $U_list[[5]]
#>            [,1]      [,2]        [,3]      [,4]      [,5]
#> [1,] 0.24211557 0.0000000 0.000000000 0.0000000 0.0000000
#> [2,] 0.08827778 0.1851573 0.000000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1919922 0.407831020 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.007134864 0.4682320 0.0000000
#> [5,] 0.00000000 0.0000000 0.000000000 0.0560379 0.9668042
#> 
#> $U_list[[6]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05373296 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02451688 0.10170820 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.05681361 0.2203045 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1642305 0.2003433 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2685748 0.9532752
#> 
#> $U_list[[7]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0608077 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.4933830 0.4604336 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2315650 0.3848140 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.4667048 0.1975581 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6653989 0.8872771
#> 
#> $U_list[[8]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.47685793 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.01248983 0.3794055 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2989584 0.4294066 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3782046 0.05694344 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.82013455 0.9702802
#> 
#> $U_list[[9]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1240807 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.2229549 0.6060821 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.0986676 0.6695799 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2749632 0.86313873 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.09419495 0.9897981
#> 
#> $U_list[[10]]
#>           [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3255609 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.3290727 0.721012443 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.006760425 0.2381221 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.5494698 0.4816385 0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.3832743 0.9188889
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    6   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    5    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    7   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    9    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    9   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    8    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    4   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>