Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfil the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE. Defaults to TRUE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

An integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

set.seed(42) #set seed for repeatability

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE, by_type = FALSE
)
#> [[1]]
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.04028499 0.00000000 3.0000000 13.0000000 13.0000000
#> [2,] 0.71907852 0.73066610 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.06797628 0.5129277  0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.2895418  0.8589953  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.1331578  0.9997924
#> 
#> [[2]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1031839 0.00000000 2.0000000 4.0000000 5.0000000
#> [2,] 0.0812552 0.53939696 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.02683477 0.6530700 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.0577064 0.4031576 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.5386195 0.9745943
#> 
#> [[3]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1060541 0.0000000 6.00000000 5.0000000 12.000000
#> [2,] 0.2992341 0.4238156 0.00000000 0.0000000  0.000000
#> [3,] 0.0000000 0.2276771 0.02692106 0.0000000  0.000000
#> [4,] 0.0000000 0.0000000 0.69200364 0.5452896  0.000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2669215  0.921095
#> 
#> [[4]]
#>            [,1]       [,2]       [,3]        [,4]       [,5]
#> [1,] 0.01338708 0.00000000 5.00000000 12.00000000 13.0000000
#> [2,] 0.46380038 0.47987890 0.00000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.03799584 0.54431954  0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.07233605  0.64886531  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.01000717  0.9298632
#> 
#> [[5]]
#>            [,1]       [,2]     [,3]      [,4]       [,5]
#> [1,] 0.03468723 0.00000000 3.000000 8.0000000 15.0000000
#> [2,] 0.13688636 0.02626718 0.000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.34994000 0.110316 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.578910 0.1255333  0.0000000
#> [5,] 0.00000000 0.00000000 0.000000 0.7370301  0.8884611
#> 
#> [[6]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04601292 0.00000000 4.00000000 4.0000000 7.0000000
#> [2,] 0.41481158 0.58800405 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.01302825 0.03951386 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.82331171 0.4383862 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.5547657 0.9998585
#> 
#> [[7]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.11547275 0.0000000 8.0000000 8.0000000 9.0000000
#> [2,] 0.09164534 0.3123506 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1387859 0.3982819 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1783419 0.1738805 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7564341 0.9780854
#> 
#> [[8]]
#>           [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1488356 0.00000000 4.00000000 9.0000000 9.0000000
#> [2,] 0.2305748 0.39486112 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.03403526 0.61030616 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.05471228 0.1821819 0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.6707913 0.8563107
#> 
#> [[9]]
#>           [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.3743886 0.00000000 3.0000000 9.0000000 13.0000000
#> [2,] 0.2859382 0.79037161 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.09338711 0.5037837 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.3851413 0.6939507  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.3009825  0.9964883
#> 
#> [[10]]
#>           [,1]      [,2]      [,3]       [,4]     [,5]
#> [1,] 0.2734565 0.0000000 7.0000000 6.00000000 6.000000
#> [2,] 0.4519310 0.2129659 0.0000000 0.00000000 0.000000
#> [3,] 0.0000000 0.5139000 0.4078221 0.00000000 0.000000
#> [4,] 0.0000000 0.0000000 0.4155600 0.01237967 0.000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.98746336 0.999866
#> 

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(
  fun = "lambda", arg = NA, lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]       [,2]        [,3]      [,4]      [,5]
#> [1,] 0.22918111 0.00000000 1.000000000 6.0000000 8.0000000
#> [2,] 0.04399029 0.07681899 0.000000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.42764990 0.526576143 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.004637456 0.7146653 0.0000000
#> [5,] 0.00000000 0.00000000 0.000000000 0.2172461 0.9351928
#> 
#> $A_list[[2]]
#>           [,1]       [,2]       [,3]       [,4]     [,5]
#> [1,] 0.0297293 0.00000000 2.00000000 10.0000000 6.000000
#> [2,] 0.1663619 0.08967374 0.00000000  0.0000000 0.000000
#> [3,] 0.0000000 0.37218398 0.51560726  0.0000000 0.000000
#> [4,] 0.0000000 0.00000000 0.05938074  0.1153077 0.000000
#> [5,] 0.0000000 0.00000000 0.00000000  0.6563415 0.905483
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08244935 0.00000000 7.0000000 8.0000000 5.0000000
#> [2,] 0.15943102 0.55226060 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.07754939 0.2277318 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.5488019 0.7305917 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2410545 0.9741093
#> 
#> $A_list[[4]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2788292 0.0000000 1.00000000 5.0000000 13.0000000
#> [2,] 0.0816387 0.1000315 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.4670093 0.55005830 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.08532858 0.4420239  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.5205884  0.9806803
#> 
#> $A_list[[5]]
#>           [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1325254 0.00000000 4.00000000 9.0000000 13.0000000
#> [2,] 0.3033947 0.47239983 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.02472252 0.08982024 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.54480458 0.4485873  0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.4844810  0.9572251
#> 
#> $A_list[[6]]
#>             [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.006123476 0.00000000 2.00000000 8.0000000 11.0000000
#> [2,] 0.115135471 0.56330702 0.00000000 0.0000000  0.0000000
#> [3,] 0.000000000 0.01036286 0.02123003 0.0000000  0.0000000
#> [4,] 0.000000000 0.00000000 0.61531368 0.2875341  0.0000000
#> [5,] 0.000000000 0.00000000 0.00000000 0.5641554  0.8894393
#> 
#> $A_list[[7]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.32910762 0.0000000 4.0000000 8.00000000 12.0000000
#> [2,] 0.09854941 0.3338640 0.0000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.1485169 0.1946671 0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3447584 0.02337579  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.76059019  0.9157293
#> 
#> $A_list[[8]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.20627958 0.0000000 2.0000000 4.0000000 12.0000000
#> [2,] 0.06736827 0.4074833 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1739123 0.6007670 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1485543 0.6497622  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1988274  0.8510634
#> 
#> $A_list[[9]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.3228405 0.0000000 1.0000000 3.0000000 11.0000000
#> [2,] 0.2239105 0.3348644 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.2150442 0.4400701 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.2492552 0.5668731  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2118358  0.8872958
#> 
#> $A_list[[10]]
#>           [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3064879 0.000000000 4.0000000 8.0000000 9.0000000
#> [2,] 0.3251220 0.715547878 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.004424139 0.4181208 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.4622338 0.7919444 0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.1587100 0.9538969
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]       [,2]        [,3]      [,4]      [,5]
#> [1,] 0.22918111 0.00000000 0.000000000 0.0000000 0.0000000
#> [2,] 0.04399029 0.07681899 0.000000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.42764990 0.526576143 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.004637456 0.7146653 0.0000000
#> [5,] 0.00000000 0.00000000 0.000000000 0.2172461 0.9351928
#> 
#> $U_list[[2]]
#>           [,1]       [,2]       [,3]      [,4]     [,5]
#> [1,] 0.0297293 0.00000000 0.00000000 0.0000000 0.000000
#> [2,] 0.1663619 0.08967374 0.00000000 0.0000000 0.000000
#> [3,] 0.0000000 0.37218398 0.51560726 0.0000000 0.000000
#> [4,] 0.0000000 0.00000000 0.05938074 0.1153077 0.000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.6563415 0.905483
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08244935 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.15943102 0.55226060 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.07754939 0.2277318 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.5488019 0.7305917 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2410545 0.9741093
#> 
#> $U_list[[4]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2788292 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.0816387 0.1000315 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.4670093 0.55005830 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.08532858 0.4420239 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.5205884 0.9806803
#> 
#> $U_list[[5]]
#>           [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1325254 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.3033947 0.47239983 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.02472252 0.08982024 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.54480458 0.4485873 0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.4844810 0.9572251
#> 
#> $U_list[[6]]
#>             [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.006123476 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.115135471 0.56330702 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.01036286 0.02123003 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.61531368 0.2875341 0.0000000
#> [5,] 0.000000000 0.00000000 0.00000000 0.5641554 0.8894393
#> 
#> $U_list[[7]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.32910762 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.09854941 0.3338640 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1485169 0.1946671 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3447584 0.02337579 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.76059019 0.9157293
#> 
#> $U_list[[8]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.20627958 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.06736827 0.4074833 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1739123 0.6007670 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1485543 0.6497622 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1988274 0.8510634
#> 
#> $U_list[[9]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3228405 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2239105 0.3348644 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2150442 0.4400701 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2492552 0.5668731 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2118358 0.8872958
#> 
#> $U_list[[10]]
#>           [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3064879 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.3251220 0.715547878 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.004424139 0.4181208 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.4622338 0.7919444 0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.1587100 0.9538969
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    6    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   10    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    8    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    5   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    9   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    8   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    4   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    3   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(
  fun = "eigs", arg = "lambda", lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>             [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.152933162 0.00000000 4.00000000 12.0000000 12.0000000
#> [2,] 0.002458585 0.23643444 0.00000000  0.0000000  0.0000000
#> [3,] 0.000000000 0.07509351 0.06410022  0.0000000  0.0000000
#> [4,] 0.000000000 0.00000000 0.34899847  0.3696469  0.0000000
#> [5,] 0.000000000 0.00000000 0.00000000  0.0619831  0.9368801
#> 
#> $A_list[[2]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.09494256 0.00000000 5.00000000 10.0000000 8.0000000
#> [2,] 0.18978251 0.01904264 0.00000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.39555674 0.42954062  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.07203381  0.4376712 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.2536738 0.8036641
#> 
#> $A_list[[3]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.18644645 0.0000000 0.00000000 9.00000000 9.0000000
#> [2,] 0.09038931 0.1087953 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.6596983 0.72585376 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.06387756 0.08432687 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.78343103 0.9856884
#> 
#> $A_list[[4]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.45729467 0.000000000 6.0000000 9.0000000 5.0000000
#> [2,] 0.03343593 0.003165349 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.493594716 0.4413990 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.2796051 0.6012603 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.1536419 0.9049761
#> 
#> $A_list[[5]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3858121 0.0000000 2.0000000 4.0000000 9.0000000
#> [2,] 0.2123301 0.4373879 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1862342 0.4867804 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1872237 0.2695684 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4209283 0.6966858
#> 
#> $A_list[[6]]
#>             [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.036019608 0.0000000 3.0000000 10.0000000 12.0000000
#> [2,] 0.007165058 0.1339785 0.0000000  0.0000000  0.0000000
#> [3,] 0.000000000 0.4912836 0.7179049  0.0000000  0.0000000
#> [4,] 0.000000000 0.0000000 0.1357687  0.2792170  0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000  0.6344899  0.9277227
#> 
#> $A_list[[7]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1203353 0.00000000 7.0000000 9.00000000 12.0000000
#> [2,] 0.1888705 0.50832213 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.06529745 0.3536676 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.2470648 0.73727607  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.06073812  0.8385156
#> 
#> $A_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.02237988 0.00000000 3.0000000 6.0000000 6.0000000
#> [2,] 0.28048090 0.03421738 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.27905428 0.1497138 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3167623 0.4295758 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2502725 0.7657977
#> 
#> $A_list[[9]]
#>            [,1]        [,2]      [,3]        [,4]      [,5]
#> [1,] 0.03170976 0.000000000 4.0000000 12.00000000 8.0000000
#> [2,] 0.07241375 0.510524515 0.0000000  0.00000000 0.0000000
#> [3,] 0.00000000 0.002057245 0.4482105  0.00000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.1720814  0.04502507 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000  0.65478296 0.9373769
#> 
#> $A_list[[10]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.17741278 0.0000000 3.0000000 13.0000000 15.0000000
#> [2,] 0.04512856 0.0425828 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.4463999 0.2472488  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2645892  0.1960029  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.4650027  0.9723187
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>             [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.152933162 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.002458585 0.23643444 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.07509351 0.06410022 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.34899847 0.3696469 0.0000000
#> [5,] 0.000000000 0.00000000 0.00000000 0.0619831 0.9368801
#> 
#> $U_list[[2]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.09494256 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.18978251 0.01904264 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.39555674 0.42954062 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.07203381 0.4376712 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.2536738 0.8036641
#> 
#> $U_list[[3]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.18644645 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.09038931 0.1087953 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.6596983 0.72585376 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.06387756 0.08432687 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.78343103 0.9856884
#> 
#> $U_list[[4]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.45729467 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03343593 0.003165349 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.493594716 0.4413990 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.2796051 0.6012603 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.1536419 0.9049761
#> 
#> $U_list[[5]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3858121 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2123301 0.4373879 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1862342 0.4867804 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1872237 0.2695684 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4209283 0.6966858
#> 
#> $U_list[[6]]
#>             [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.036019608 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.007165058 0.1339785 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.4912836 0.7179049 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.1357687 0.2792170 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.6344899 0.9277227
#> 
#> $U_list[[7]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1203353 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.1888705 0.50832213 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.06529745 0.3536676 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.2470648 0.73727607 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.06073812 0.8385156
#> 
#> $U_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.02237988 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.28048090 0.03421738 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.27905428 0.1497138 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3167623 0.4295758 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2502725 0.7657977
#> 
#> $U_list[[9]]
#>            [,1]        [,2]      [,3]       [,4]      [,5]
#> [1,] 0.03170976 0.000000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.07241375 0.510524515 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.002057245 0.4482105 0.00000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.1720814 0.04502507 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.65478296 0.9373769
#> 
#> $U_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.17741278 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04512856 0.0425828 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.4463999 0.2472488 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2645892 0.1960029 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4650027 0.9723187
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   12   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5   10    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    9    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    9    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    4    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    9   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    6    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   12    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   13   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,] 0.32261745 0.0000000 7.00000000 14.0000000 13.0000000
#> [2,] 0.09501365 0.3978349 0.00000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.1884375 0.01659362  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.71862085  0.4833590  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000  0.3742096  0.9073485
#> 
#> $A_list[[2]]
#>            [,1]       [,2]      [,3]        [,4]       [,5]
#> [1,] 0.08612368 0.00000000 6.0000000 10.00000000 15.0000000
#> [2,] 0.04560208 0.06975515 0.0000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.28447678 0.1275618  0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.7625399  0.08617495  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.81587337  0.9529402
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.03211654 0.00000000 3.0000000 9.00000000 10.0000000
#> [2,] 0.40177260 0.38269030 0.0000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.05728678 0.3170391 0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.1289007 0.02878054  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.49694707  0.7991334
#> 
#> $A_list[[4]]
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.10414152 0.00000000 3.00000000 10.0000000 14.0000000
#> [2,] 0.04989928 0.02917173 0.00000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.33138597 0.45498595  0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.02627893  0.6504826  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.2661919  0.9764586
#> 
#> $A_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.21091154 0.0000000 4.0000000 7.0000000 6.0000000
#> [2,] 0.04430196 0.1218693 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1671190 0.3703027 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1104442 0.1853642 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3469549 0.9590821
#> 
#> $A_list[[6]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07181072 0.00000000 4.0000000 9.0000000 8.0000000
#> [2,] 0.38947376 0.38994639 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.09278298 0.4366518 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1616440 0.6035881 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1807365 0.8738491
#> 
#> $A_list[[7]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1009137 0.0000000 0.0000000 7.0000000 6.0000000
#> [2,] 0.0431711 0.1573180 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.6711804 0.6288931 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2376405 0.1273381 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.7859082 0.9295329
#> 
#> $A_list[[8]]
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.07605220 0.00000000 9.00000000 10.0000000 10.0000000
#> [2,] 0.07720887 0.02650316 0.00000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.22442566 0.06417721  0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.36027570  0.3518009  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.4589660  0.8476694
#> 
#> $A_list[[9]]
#>          [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.490599 0.0000000 3.00000000 12.0000000 7.0000000
#> [2,] 0.241938 0.7506678 0.00000000  0.0000000 0.0000000
#> [3,] 0.000000 0.1099322 0.86236715  0.0000000 0.0000000
#> [4,] 0.000000 0.0000000 0.01964878  0.6312540 0.0000000
#> [5,] 0.000000 0.0000000 0.00000000  0.3142358 0.9486546
#> 
#> $A_list[[10]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.5799961 0.0000000 8.0000000 9.0000000 17.0000000
#> [2,] 0.1290812 0.4113872 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.3183108 0.5891245 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.1415227 0.4881498  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.3423692  0.8758764
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.32261745 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.09501365 0.3978349 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1884375 0.01659362 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.71862085 0.4833590 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3742096 0.9073485
#> 
#> $U_list[[2]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.08612368 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.04560208 0.06975515 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.28447678 0.1275618 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.7625399 0.08617495 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.81587337 0.9529402
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.03211654 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.40177260 0.38269030 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.05728678 0.3170391 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1289007 0.02878054 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.49694707 0.7991334
#> 
#> $U_list[[4]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.10414152 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.04989928 0.02917173 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.33138597 0.45498595 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.02627893 0.6504826 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.2661919 0.9764586
#> 
#> $U_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.21091154 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04430196 0.1218693 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1671190 0.3703027 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1104442 0.1853642 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3469549 0.9590821
#> 
#> $U_list[[6]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07181072 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.38947376 0.38994639 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.09278298 0.4366518 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1616440 0.6035881 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1807365 0.8738491
#> 
#> $U_list[[7]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1009137 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.0431711 0.1573180 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.6711804 0.6288931 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2376405 0.1273381 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.7859082 0.9295329
#> 
#> $U_list[[8]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.07605220 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.07720887 0.02650316 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.22442566 0.06417721 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.36027570 0.3518009 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.4589660 0.8476694
#> 
#> $U_list[[9]]
#>          [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.490599 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.241938 0.7506678 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000 0.1099322 0.86236715 0.0000000 0.0000000
#> [4,] 0.000000 0.0000000 0.01964878 0.6312540 0.0000000
#> [5,] 0.000000 0.0000000 0.00000000 0.3142358 0.9486546
#> 
#> $U_list[[10]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.5799961 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1290812 0.4113872 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3183108 0.5891245 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1415227 0.4881498 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.3423692 0.8758764
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7   14   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   10   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    7    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    9    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    7    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    9   10   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   12    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8    9   17
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>