Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable lambda values. The function attempts to find matrices that fulfill the lambda criteria, discarding unacceptable matrices. If it takes more than 1000 attempts to find a suitable matrix model, then an error is produced.

Usage

generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

lower_lambda

The lower bound for acceptable lambda values. Default is 0.9.

upper_lambda

The upper bound for acceptable lambda values. Default is 1.1.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional list specifying (1) a function that outputs a metric derived from an A matrix and (2) acceptable bounds that are used to constrain matrix outputs. E.g. list("generation.time", c(5,10)) to specify that generate_mpm_set should only output MPMs with a generation time between 5 and 10. Default is NULL.

attempts

an integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)
#> $A_list
#> $A_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.20099990 0.00000000 5.0000000 6.0000000 10.000000
#> [2,] 0.02863383 0.36344661 0.0000000 0.0000000  0.000000
#> [3,] 0.00000000 0.08215543 0.0179438 0.0000000  0.000000
#> [4,] 0.00000000 0.00000000 0.4971116 0.1264745  0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6776361  0.983769
#> 
#> $A_list[[2]]
#>           [,1]        [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2756682 0.000000000 6.00000000 7.0000000 16.0000000
#> [2,] 0.1476305 0.546073267 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.009984465 0.75743679 0.0000000  0.0000000
#> [4,] 0.0000000 0.000000000 0.07976745 0.0878548  0.0000000
#> [5,] 0.0000000 0.000000000 0.00000000 0.7848476  0.9467092
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.29639051 0.0000000 2.0000000 9.0000000 15.0000000
#> [2,] 0.01556814 0.2855882 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.4209131 0.1750581 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.5562840 0.3767262  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4679580  0.9875311
#> 
#> $A_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.06955233 0.0000000 4.0000000 5.0000000 10.0000000
#> [2,] 0.07907788 0.3056856 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1865541 0.1029445 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.4798498 0.6247633  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2196091  0.8807173
#> 
#> $A_list[[5]]
#>           [,1]      [,2]        [,3]      [,4]       [,5]
#> [1,] 0.2573025 0.0000000 8.000000000 5.0000000 11.0000000
#> [2,] 0.1406432 0.3793544 0.000000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1326322 0.006541837 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.526898758 0.6390245  0.0000000
#> [5,] 0.0000000 0.0000000 0.000000000 0.1547418  0.9439308
#> 
#> $A_list[[6]]
#>             [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.297574631 0.0000000 5.0000000 8.0000000 8.0000000
#> [2,] 0.006518948 0.4209899 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.2898602 0.5753706 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.1581245 0.1624531 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.7029352 0.9665908
#> 
#> $A_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]    [,5]
#> [1,] 0.06111573 0.0000000 2.0000000 7.0000000 8.00000
#> [2,] 0.28765787 0.3603954 0.0000000 0.0000000 0.00000
#> [3,] 0.00000000 0.1665081 0.5948303 0.0000000 0.00000
#> [4,] 0.00000000 0.0000000 0.0730143 0.1135435 0.00000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6000926 0.85626
#> 
#> $A_list[[8]]
#>             [,1]      [,2]      [,3]        [,4]      [,5]
#> [1,] 0.549437331 0.0000000 5.0000000 11.00000000 5.0000000
#> [2,] 0.008614858 0.1434581 0.0000000  0.00000000 0.0000000
#> [3,] 0.000000000 0.7137644 0.3438684  0.00000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.5298967  0.88041695 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000  0.06439547 0.9559577
#> 
#> $A_list[[9]]
#>            [,1]      [,2]      [,3]     [,4]       [,5]
#> [1,] 0.41406609 0.0000000 7.0000000 7.000000 11.0000000
#> [2,] 0.03468226 0.3620542 0.0000000 0.000000  0.0000000
#> [3,] 0.00000000 0.1349057 0.4590503 0.000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1833704 0.456872  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.492844  0.9706123
#> 
#> $A_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.12521020 0.0000000 5.0000000 5.0000000 8.0000000
#> [2,] 0.04927332 0.4227989 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2490446 0.0468008 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.7243134 0.7663143 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1149266 0.9551785
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.20099990 0.00000000 0.0000000 0.0000000 0.000000
#> [2,] 0.02863383 0.36344661 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.08215543 0.0179438 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.4971116 0.1264745 0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6776361 0.983769
#> 
#> $U_list[[2]]
#>           [,1]        [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2756682 0.000000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1476305 0.546073267 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.009984465 0.75743679 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.07976745 0.0878548 0.0000000
#> [5,] 0.0000000 0.000000000 0.00000000 0.7848476 0.9467092
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.29639051 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01556814 0.2855882 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.4209131 0.1750581 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.5562840 0.3767262 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4679580 0.9875311
#> 
#> $U_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.06955233 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.07907788 0.3056856 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1865541 0.1029445 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4798498 0.6247633 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2196091 0.8807173
#> 
#> $U_list[[5]]
#>           [,1]      [,2]        [,3]      [,4]      [,5]
#> [1,] 0.2573025 0.0000000 0.000000000 0.0000000 0.0000000
#> [2,] 0.1406432 0.3793544 0.000000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1326322 0.006541837 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.526898758 0.6390245 0.0000000
#> [5,] 0.0000000 0.0000000 0.000000000 0.1547418 0.9439308
#> 
#> $U_list[[6]]
#>             [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.297574631 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.006518948 0.4209899 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.2898602 0.5753706 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.1581245 0.1624531 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.7029352 0.9665908
#> 
#> $U_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]    [,5]
#> [1,] 0.06111573 0.0000000 0.0000000 0.0000000 0.00000
#> [2,] 0.28765787 0.3603954 0.0000000 0.0000000 0.00000
#> [3,] 0.00000000 0.1665081 0.5948303 0.0000000 0.00000
#> [4,] 0.00000000 0.0000000 0.0730143 0.1135435 0.00000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6000926 0.85626
#> 
#> $U_list[[8]]
#>             [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.549437331 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.008614858 0.1434581 0.0000000 0.00000000 0.0000000
#> [3,] 0.000000000 0.7137644 0.3438684 0.00000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.5298967 0.88041695 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.06439547 0.9559577
#> 
#> $U_list[[9]]
#>            [,1]      [,2]      [,3]     [,4]      [,5]
#> [1,] 0.41406609 0.0000000 0.0000000 0.000000 0.0000000
#> [2,] 0.03468226 0.3620542 0.0000000 0.000000 0.0000000
#> [3,] 0.00000000 0.1349057 0.4590503 0.000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1833704 0.456872 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.492844 0.9706123
#> 
#> $U_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.12521020 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04927332 0.4227989 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2490446 0.0468008 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.7243134 0.7663143 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1149266 0.9551785
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    7   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    9   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    5   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8    5   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5   11    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    7   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    5    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

library(popbio)
generate_mpm_set(
  n = 1, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE,
  constraint = list("generation.time", c(40, 50))
)
#> $A_list
#> $A_list[[1]]
#>           [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1977237 0.000000000 4.0000000 5.0000000 9.0000000
#> [2,] 0.3581691 0.792737383 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.001475846 0.1323031 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.6668426 0.2268235 0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.7280901 0.9741684
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1977237 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.3581691 0.792737383 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.001475846 0.1323031 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.6668426 0.2268235 0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.7280901 0.9741684
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    5    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>