Generate lists of matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable lambda values. The function attempts to find matrices that fulfill the lambda criteria, discarding unacceptable matrices. If it takes more than 1000 attempts to find a suitable matrix model, then an error is produced.

Usage

generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE
)

Arguments

n

The number of MPMs to generate. Default is 10.

lower_lambda

The lower bound for acceptable lambda values. Default is 0.9.

upper_lambda

The upper bound for acceptable lambda values. Default is 1.1.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)
#> $A_list
#> $A_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08406612 0.00000000 4.0000000 5.0000000 8.0000000
#> [2,] 0.20669133 0.03561627 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.38767493 0.1315684 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3034208 0.2817810 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2642514 0.8791604
#> 
#> $A_list[[2]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.24630397 0.00000000 5.0000000 7.00000000 8.0000000
#> [2,] 0.09343289 0.46081310 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.05060626 0.2497827 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4191242 0.70598635 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.09712258 0.8357981
#> 
#> $A_list[[3]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.29308962 0.0000000 4.00000000 4.0000000 5.0000000
#> [2,] 0.01673647 0.4578758 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3125445 0.88656542 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.04704555 0.3063487 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.6555767 0.9949944
#> 
#> $A_list[[4]]
#>           [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1844297 0.00000000 4.0000000 9.0000000 15.0000000
#> [2,] 0.2211504 0.49760271 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.02283487 0.5030562 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.1824151 0.5326546  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.4623351  0.9977389
#> 
#> $A_list[[5]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.3356919 0.0000000 9.0000000 6.0000000 11.0000000
#> [2,] 0.0378977 0.3824298 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1764402 0.4949150 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.1767051 0.0505027  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.8449470  0.9240002
#> 
#> $A_list[[6]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2810448 0.0000000 1.00000000 4.0000000 14.0000000
#> [2,] 0.1014773 0.3400761 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1426181 0.66658369 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.04509492 0.4043196  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3509684  0.7998701
#> 
#> $A_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.12755407 0.0000000 3.0000000 8.0000000 5.0000000
#> [2,] 0.02036622 0.1692137 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.4152229 0.4614509 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1538901 0.4406283 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1879739 0.9047152
#> 
#> $A_list[[8]]
#>           [,1]       [,2]      [,3]        [,4]      [,5]
#> [1,] 0.4775918 0.00000000 3.0000000 11.00000000 9.0000000
#> [2,] 0.1203152 0.65802893 0.0000000  0.00000000 0.0000000
#> [3,] 0.0000000 0.05756145 0.1954674  0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.5363555  0.06006875 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000  0.86828675 0.9319601
#> 
#> $A_list[[9]]
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.04554726 0.00000000 0.00000000 6.0000000 19.0000000
#> [2,] 0.10676814 0.26211611 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.03230526 0.04381474 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.69626041 0.6909822  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1485400  0.9998232
#> 
#> $A_list[[10]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01909707 0.00000000 5.0000000 8.0000000 7.0000000
#> [2,] 0.47079984 0.51377946 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03859988 0.5554581 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2861327 0.7347313 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1570935 0.8936522
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08406612 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.20669133 0.03561627 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.38767493 0.1315684 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3034208 0.2817810 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2642514 0.8791604
#> 
#> $U_list[[2]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.24630397 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.09343289 0.46081310 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.05060626 0.2497827 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4191242 0.70598635 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.09712258 0.8357981
#> 
#> $U_list[[3]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.29308962 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.01673647 0.4578758 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3125445 0.88656542 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.04704555 0.3063487 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.6555767 0.9949944
#> 
#> $U_list[[4]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1844297 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2211504 0.49760271 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.02283487 0.5030562 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.1824151 0.5326546 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.4623351 0.9977389
#> 
#> $U_list[[5]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3356919 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.0378977 0.3824298 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1764402 0.4949150 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1767051 0.0505027 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.8449470 0.9240002
#> 
#> $U_list[[6]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2810448 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1014773 0.3400761 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1426181 0.66658369 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.04509492 0.4043196 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3509684 0.7998701
#> 
#> $U_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.12755407 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02036622 0.1692137 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.4152229 0.4614509 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1538901 0.4406283 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1879739 0.9047152
#> 
#> $U_list[[8]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.4775918 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.1203152 0.65802893 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.05756145 0.1954674 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.5363555 0.06006875 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.86828675 0.9319601
#> 
#> $U_list[[9]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04554726 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.10676814 0.26211611 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03230526 0.04381474 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.69626041 0.6909822 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1485400 0.9998232
#> 
#> $U_list[[10]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01909707 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.47079984 0.51377946 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03859988 0.5554581 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2861327 0.7347313 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1570935 0.8936522
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    5    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    7    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    4    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    9   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    9    6   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    4   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    8    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   11    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    6   19
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>