Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfill the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

an integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE
)
#> Error in generate_mpm_set(n = 10, n_stages = 5, fecundity = c(0, 0, 4,     8, 10), archetype = 4, split = FALSE): If split is FALSE, then by_type must also be FALSE

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(fun = "lambda", arg = NA, lower = 0.9, upper = 1.1)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.04623129 0.0000000 4.00000000 8.0000000 16.0000000
#> [2,] 0.19929781 0.1480114 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1684678 0.42341229 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.08080674 0.1012408  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.7973402  0.9008713
#> 
#> $A_list[[2]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.13086147 0.00000000 9.00000000 7.0000000 11.000000
#> [2,] 0.01219071 0.05493529 0.00000000 0.0000000  0.000000
#> [3,] 0.00000000 0.75725475 0.02553352 0.0000000  0.000000
#> [4,] 0.00000000 0.00000000 0.87313139 0.8557531  0.000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1143465  0.987545
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.21068793 0.0000000 3.0000000 9.0000000 15.0000000
#> [2,] 0.03172115 0.1572462 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1171225 0.3455803 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.4638089 0.5081574  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3294787  0.8941699
#> 
#> $A_list[[4]]
#>           [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2480688 0.00000000 5.0000000 6.0000000 17.0000000
#> [2,] 0.0134266 0.38286833 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.02768124 0.5070301 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.1442138 0.5293036  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.3614421  0.9411137
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.59159000 0.00000000 5.0000000 8.0000000 5.0000000
#> [2,] 0.01712411 0.78907512 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02265376 0.7528609 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1448417 0.3424663 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6090626 0.9606187
#> 
#> $A_list[[6]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.04042515 0.0000000 2.0000000 14.0000000 8.0000000
#> [2,] 0.04929932 0.3753847 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.2287047 0.1574686  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.5686369  0.1152859 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.7865335 0.9764925
#> 
#> $A_list[[7]]
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.4701780 0.0000000 3.0000000 6.00000000 13.0000000
#> [2,] 0.0147524 0.2887703 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.2875623 0.5129822 0.00000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3275484 0.89942654  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.03712489  0.9826925
#> 
#> $A_list[[8]]
#>            [,1]        [,2]       [,3]      [,4]      [,5]
#> [1,] 0.02695663 0.000000000 2.00000000 5.0000000 6.0000000
#> [2,] 0.06075092 0.500958165 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.007562979 0.73231903 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.09446381 0.7529101 0.0000000
#> [5,] 0.00000000 0.000000000 0.00000000 0.1593567 0.9145223
#> 
#> $A_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.06525775 0.00000000 4.0000000 6.0000000 5.0000000
#> [2,] 0.22453711 0.58045257 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02778435 0.4643570 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2114333 0.3529411 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6017199 0.9647116
#> 
#> $A_list[[10]]
#>             [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.007985743 0.00000000 1.0000000 10.0000000 8.0000000
#> [2,] 0.075822136 0.22915729 0.0000000  0.0000000 0.0000000
#> [3,] 0.000000000 0.03416866 0.5256118  0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.3421395  0.1794401 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000  0.7172544 0.9380857
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04623129 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.19929781 0.1480114 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1684678 0.42341229 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.08080674 0.1012408 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.7973402 0.9008713
#> 
#> $U_list[[2]]
#>            [,1]       [,2]       [,3]      [,4]     [,5]
#> [1,] 0.13086147 0.00000000 0.00000000 0.0000000 0.000000
#> [2,] 0.01219071 0.05493529 0.00000000 0.0000000 0.000000
#> [3,] 0.00000000 0.75725475 0.02553352 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.87313139 0.8557531 0.000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1143465 0.987545
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.21068793 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03172115 0.1572462 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1171225 0.3455803 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4638089 0.5081574 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3294787 0.8941699
#> 
#> $U_list[[4]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2480688 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.0134266 0.38286833 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.02768124 0.5070301 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.1442138 0.5293036 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.3614421 0.9411137
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.59159000 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01712411 0.78907512 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02265376 0.7528609 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1448417 0.3424663 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6090626 0.9606187
#> 
#> $U_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.04042515 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04929932 0.3753847 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2287047 0.1574686 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.5686369 0.1152859 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7865335 0.9764925
#> 
#> $U_list[[7]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.4701780 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.0147524 0.2887703 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.2875623 0.5129822 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3275484 0.89942654 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.03712489 0.9826925
#> 
#> $U_list[[8]]
#>            [,1]        [,2]       [,3]      [,4]      [,5]
#> [1,] 0.02695663 0.000000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.06075092 0.500958165 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.007562979 0.73231903 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.09446381 0.7529101 0.0000000
#> [5,] 0.00000000 0.000000000 0.00000000 0.1593567 0.9145223
#> 
#> $U_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.06525775 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.22453711 0.58045257 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02778435 0.4643570 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2114333 0.3529411 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6017199 0.9647116
#> 
#> $U_list[[10]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.007985743 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.075822136 0.22915729 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.03416866 0.5256118 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.3421395 0.1794401 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.7172544 0.9380857
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    9    7   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6   17
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   14    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    6   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    5    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    6    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1   10    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(fun = "eigs", arg = "lambda", lower = 0.9, upper = 1.1)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04022567 0.00000000 6.00000000 7.0000000 8.0000000
#> [2,] 0.05569933 0.39780223 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.01956637 0.08576079 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.61355855 0.7235467 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1468692 0.9683902
#> 
#> $A_list[[2]]
#>           [,1]      [,2]       [,3]         [,4]       [,5]
#> [1,] 0.1499998 0.0000000 1.00000000 10.000000000 11.0000000
#> [2,] 0.2990411 0.2188090 0.00000000  0.000000000  0.0000000
#> [3,] 0.0000000 0.2402426 0.66732232  0.000000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.07532487  0.007785977  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000  0.838917474  0.8948816
#> 
#> $A_list[[3]]
#>            [,1]        [,2]      [,3]       [,4]      [,5]
#> [1,] 0.00588685 0.000000000 5.0000000 10.0000000 4.0000000
#> [2,] 0.25915909 0.413169946 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.002680835 0.7550967  0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.0996027  0.7406453 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000  0.1943904 0.9541706
#> 
#> $A_list[[4]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.4031156 0.0000000 7.0000000 6.0000000 12.0000000
#> [2,] 0.1297168 0.5087997 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.0372987 0.3468743 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3754735 0.2904621  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.5406643  0.8547606
#> 
#> $A_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.43785177 0.0000000 1.0000000 3.0000000 16.0000000
#> [2,] 0.03580066 0.1963211 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3447377 0.4369221 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3800737 0.3523054  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5481361  0.9377487
#> 
#> $A_list[[6]]
#>            [,1]      [,2]      [,3]        [,4]       [,5]
#> [1,] 0.01510299 0.0000000 0.0000000 10.00000000 11.0000000
#> [2,] 0.11463396 0.2179490 0.0000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.3463509 0.7119329  0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1366192  0.80400726  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.05335479  0.9636138
#> 
#> $A_list[[7]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.04104083 0.00000000 3.0000000 10.0000000 14.000000
#> [2,] 0.26089428 0.42885482 0.0000000  0.0000000  0.000000
#> [3,] 0.00000000 0.03212227 0.4722691  0.0000000  0.000000
#> [4,] 0.00000000 0.00000000 0.3878985  0.7905673  0.000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.1391503  0.938933
#> 
#> $A_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2945894 0.0000000 4.0000000 2.0000000 9.0000000
#> [2,] 0.0854030 0.5722961 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1175227 0.4651846 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3396317 0.3533088 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4981075 0.9124304
#> 
#> $A_list[[9]]
#>            [,1]       [,2]       [,3]       [,4]     [,5]
#> [1,] 0.40457373 0.00000000 2.00000000 10.0000000 6.000000
#> [2,] 0.09274162 0.50944817 0.00000000  0.0000000 0.000000
#> [3,] 0.00000000 0.04366773 0.78959029  0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.06640043  0.5080241 0.000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.4203905 0.976001
#> 
#> $A_list[[10]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.40585130 0.0000000 5.0000000 7.00000000 10.0000000
#> [2,] 0.02353008 0.2697922 0.0000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.4720871 0.6197566 0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2103076 0.81939246  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.04988938  0.9536784
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04022567 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.05569933 0.39780223 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.01956637 0.08576079 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.61355855 0.7235467 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1468692 0.9683902
#> 
#> $U_list[[2]]
#>           [,1]      [,2]       [,3]        [,4]      [,5]
#> [1,] 0.1499998 0.0000000 0.00000000 0.000000000 0.0000000
#> [2,] 0.2990411 0.2188090 0.00000000 0.000000000 0.0000000
#> [3,] 0.0000000 0.2402426 0.66732232 0.000000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.07532487 0.007785977 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.838917474 0.8948816
#> 
#> $U_list[[3]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.00588685 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.25915909 0.413169946 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.002680835 0.7550967 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.0996027 0.7406453 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.1943904 0.9541706
#> 
#> $U_list[[4]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.4031156 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1297168 0.5087997 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.0372987 0.3468743 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3754735 0.2904621 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.5406643 0.8547606
#> 
#> $U_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.43785177 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03580066 0.1963211 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3447377 0.4369221 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3800737 0.3523054 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5481361 0.9377487
#> 
#> $U_list[[6]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.01510299 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.11463396 0.2179490 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.3463509 0.7119329 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1366192 0.80400726 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.05335479 0.9636138
#> 
#> $U_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.04104083 0.00000000 0.0000000 0.0000000 0.000000
#> [2,] 0.26089428 0.42885482 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.03212227 0.4722691 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.3878985 0.7905673 0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1391503 0.938933
#> 
#> $U_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2945894 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.0854030 0.5722961 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1175227 0.4651846 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3396317 0.3533088 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4981075 0.9124304
#> 
#> $U_list[[9]]
#>            [,1]       [,2]       [,3]      [,4]     [,5]
#> [1,] 0.40457373 0.00000000 0.00000000 0.0000000 0.000000
#> [2,] 0.09274162 0.50944817 0.00000000 0.0000000 0.000000
#> [3,] 0.00000000 0.04366773 0.78959029 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.06640043 0.5080241 0.000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.4203905 0.976001
#> 
#> $U_list[[10]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.40585130 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.02353008 0.2697922 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.4720871 0.6197566 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2103076 0.81939246 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.04988938 0.9536784
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    7    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1   10   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5   10    4
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    6   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    3   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0   10   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    2    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   10    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    7   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>           [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.4940157 0.0000000 6.00000000 11.0000000 5.0000000
#> [2,] 0.1440595 0.5470211 0.00000000  0.0000000 0.0000000
#> [3,] 0.0000000 0.1433050 0.65087086  0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.09796024  0.1464969 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000  0.7784453 0.9895158
#> 
#> $A_list[[2]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04920269 0.00000000 1.00000000 4.0000000 8.0000000
#> [2,] 0.34027876 0.59150288 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.05422741 0.08028746 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.66602077 0.2639009 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6098982 0.9890875
#> 
#> $A_list[[3]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1824348 0.0000000 3.0000000 7.0000000 18.0000000
#> [2,] 0.3706747 0.6303456 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1197402 0.1497223 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.6340007 0.3443214  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6084378  0.9690089
#> 
#> $A_list[[4]]
#>           [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.2269847 0.0000000 4.00000000 11.0000000 6.0000000
#> [2,] 0.0499557 0.4718554 0.00000000  0.0000000 0.0000000
#> [3,] 0.0000000 0.2558106 0.05602945  0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.74797987  0.4241585 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000  0.4825079 0.9178029
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.44003163 0.00000000 5.0000000 6.0000000 8.0000000
#> [2,] 0.06785057 0.65421770 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.04986438 0.1644860 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.7395534 0.7347958 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2323565 0.9830788
#> 
#> $A_list[[6]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1348258 0.0000000 6.0000000 9.0000000 16.0000000
#> [2,] 0.1027750 0.2728379 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.3550627 0.6054541 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.1651691 0.5074385  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2637126  0.8807161
#> 
#> $A_list[[7]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.09153374 0.00000000 6.00000000 11.0000000 7.0000000
#> [2,] 0.21148883 0.53627134 0.00000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.04216482 0.67831072  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.08776629  0.5160776 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.4148551 0.9684819
#> 
#> $A_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1891462 0.0000000 4.0000000 8.0000000 8.0000000
#> [2,] 0.1449615 0.1009245 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.4739360 0.2416487 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3885797 0.3918694 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4433553 0.9679446
#> 
#> $A_list[[9]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1736104 0.0000000 4.00000000 8.0000000 8.0000000
#> [2,] 0.4344886 0.2712913 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3963148 0.06636123 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.64178082 0.6374186 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2154958 0.8673134
#> 
#> $A_list[[10]]
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.5095418 0.0000000 6.0000000 6.00000000 14.0000000
#> [2,] 0.2546204 0.4016969 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.3896682 0.6986541 0.00000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.1221957 0.09547914  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.88167082  0.9812988
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.4940157 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1440595 0.5470211 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1433050 0.65087086 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.09796024 0.1464969 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.7784453 0.9895158
#> 
#> $U_list[[2]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04920269 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.34027876 0.59150288 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.05422741 0.08028746 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.66602077 0.2639009 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6098982 0.9890875
#> 
#> $U_list[[3]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1824348 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.3706747 0.6303456 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1197402 0.1497223 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.6340007 0.3443214 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6084378 0.9690089
#> 
#> $U_list[[4]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2269847 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.0499557 0.4718554 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2558106 0.05602945 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.74797987 0.4241585 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.4825079 0.9178029
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.44003163 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.06785057 0.65421770 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.04986438 0.1644860 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.7395534 0.7347958 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2323565 0.9830788
#> 
#> $U_list[[6]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1348258 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1027750 0.2728379 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3550627 0.6054541 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1651691 0.5074385 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2637126 0.8807161
#> 
#> $U_list[[7]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.09153374 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.21148883 0.53627134 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.04216482 0.67831072 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.08776629 0.5160776 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.4148551 0.9684819
#> 
#> $U_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1891462 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1449615 0.1009245 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.4739360 0.2416487 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3885797 0.3918694 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4433553 0.9679446
#> 
#> $U_list[[9]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1736104 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.4344886 0.2712913 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3963148 0.06636123 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.64178082 0.6374186 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2154958 0.8673134
#> 
#> $U_list[[10]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.5095418 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.2546204 0.4016969 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.3896682 0.6986541 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1221957 0.09547914 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.88167082 0.9812988
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   11    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    4    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    7   18
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   11    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    9   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   11    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    6   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>