Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfil the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE. Defaults to TRUE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

An integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE, by_type = FALSE
)
#> [[1]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.07973508 0.0000000 7.0000000 7.0000000 12.0000000
#> [2,] 0.07626266 0.2758943 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3137862 0.5936916 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2800675 0.2534801  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7263330  0.9845833
#> 
#> [[2]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1457750 0.00000000 2.0000000 7.0000000 8.0000000
#> [2,] 0.1995919 0.09299923 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.46525889 0.6046724 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.1402542 0.3026146 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.5004493 0.9325951
#> 
#> [[3]]
#>           [,1]       [,2]      [,3]     [,4]       [,5]
#> [1,] 0.1383285 0.00000000 0.0000000 6.000000 12.0000000
#> [2,] 0.1298041 0.25886093 0.0000000 0.000000  0.0000000
#> [3,] 0.0000000 0.01149898 0.2461773 0.000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.4046551 0.615292  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.188382  0.9173377
#> 
#> [[4]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1559894 0.0000000 1.0000000 5.0000000 10.0000000
#> [2,] 0.6071385 0.4510531 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.3209724 0.6726836 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.1392090 0.6989555  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2185933  0.9286816
#> 
#> [[5]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.05415519 0.0000000 4.00000000 3.00000000 9.0000000
#> [2,] 0.15942383 0.2404292 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1693787 0.03245713 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.71888209 0.92606553 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.01660534 0.9449243
#> 
#> [[6]]
#>           [,1]        [,2]       [,3]         [,4]      [,5]
#> [1,] 0.2506707 0.000000000 4.00000000 12.000000000 11.000000
#> [2,] 0.1165497 0.003236457 0.00000000  0.000000000  0.000000
#> [3,] 0.0000000 0.460256563 0.04666278  0.000000000  0.000000
#> [4,] 0.0000000 0.000000000 0.86699134  0.945971314  0.000000
#> [5,] 0.0000000 0.000000000 0.00000000  0.002201134  0.959957
#> 
#> [[7]]
#>              [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.0004514112 0.0000000 4.0000000 5.0000000 10.0000000
#> [2,] 0.1113845815 0.2227474 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000000 0.4269933 0.5083870 0.0000000  0.0000000
#> [4,] 0.0000000000 0.0000000 0.1852437 0.5977088  0.0000000
#> [5,] 0.0000000000 0.0000000 0.0000000 0.2930168  0.9662778
#> 
#> [[8]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.06107285 0.0000000 3.0000000 6.0000000 11.0000000
#> [2,] 0.15982350 0.2261292 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3820153 0.1433044 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.6551849 0.6390876  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2516576  0.9105182
#> 
#> [[9]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.10904188 0.0000000 4.00000000 8.0000000 11.0000000
#> [2,] 0.08543115 0.2792185 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1114829 0.03123352 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.69641829 0.4986337  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.4879098  0.9931576
#> 
#> [[10]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1090197 0.0000000 5.0000000 3.0000000 16.0000000
#> [2,] 0.2960445 0.3860057 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.2560240 0.5130864 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.1559048 0.1775637  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6344341  0.9542505
#> 

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(
  fun = "lambda", arg = NA, lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.01713022 0.0000000 6.0000000 6.0000000 14.0000000
#> [2,] 0.20982748 0.3979043 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1020806 0.2779488 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2885967 0.3524984  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2317995  0.6035654
#> 
#> $A_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.28494128 0.0000000 7.0000000 9.0000000 12.0000000
#> [2,] 0.03941275 0.1210972 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2836815 0.2006138 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3202549 0.6413552  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1107985  0.9279116
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.47518854 0.0000000 4.0000000 6.00000000 7.0000000
#> [2,] 0.02140806 0.3130279 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.3123082 0.1595701 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.7625623 0.93950291 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.01369407 0.9580003
#> 
#> $A_list[[4]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.24884415 0.00000000 1.0000000 6.0000000 10.0000000
#> [2,] 0.02917995 0.44696378 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.05773925 0.3658179 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.1864414 0.3662930  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5384299  0.9367497
#> 
#> $A_list[[5]]
#>             [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.474365587 0.00000000 6.0000000 10.0000000 8.0000000
#> [2,] 0.005938972 0.60692355 0.0000000  0.0000000 0.0000000
#> [3,] 0.000000000 0.04882422 0.1012624  0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.7498044  0.5006243 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000  0.4580697 0.9871238
#> 
#> $A_list[[6]]
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1008478 0.0000000 2.0000000 7.00000000 10.0000000
#> [2,] 0.2590016 0.1816724 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.2521469 0.1469994 0.00000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3265011 0.77987985  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.01004787  0.9016346
#> 
#> $A_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.04988646 0.00000000 6.0000000 8.0000000 14.0000000
#> [2,] 0.01872387 0.54972128 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.04052471 0.0267861 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.9580687 0.1579888  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.8313010  0.9896272
#> 
#> $A_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.17032236 0.00000000 6.0000000 4.0000000 7.0000000
#> [2,] 0.01597962 0.07635881 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.17328007 0.2165077 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4206939 0.5173125 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1731996 0.9374756
#> 
#> $A_list[[9]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.43326880 0.0000000 9.0000000 13.0000000 16.0000000
#> [2,] 0.01528773 0.2057601 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.3021501 0.3679075  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2604911  0.5171856  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.2263112  0.9787172
#> 
#> $A_list[[10]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.06145321 0.00000000 1.0000000 7.0000000 8.0000000
#> [2,] 0.31900255 0.44396571 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.01614824 0.3348477 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1787952 0.2675066 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5052584 0.9120597
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01713022 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.20982748 0.3979043 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1020806 0.2779488 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2885967 0.3524984 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2317995 0.6035654
#> 
#> $U_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.28494128 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03941275 0.1210972 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2836815 0.2006138 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3202549 0.6413552 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1107985 0.9279116
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.47518854 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.02140806 0.3130279 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.3123082 0.1595701 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.7625623 0.93950291 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.01369407 0.9580003
#> 
#> $U_list[[4]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.24884415 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02917995 0.44696378 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.05773925 0.3658179 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1864414 0.3662930 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5384299 0.9367497
#> 
#> $U_list[[5]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.474365587 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.005938972 0.60692355 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.04882422 0.1012624 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.7498044 0.5006243 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.4580697 0.9871238
#> 
#> $U_list[[6]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1008478 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.2590016 0.1816724 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.2521469 0.1469994 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3265011 0.77987985 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.01004787 0.9016346
#> 
#> $U_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.04988646 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01872387 0.54972128 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.04052471 0.0267861 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.9580687 0.1579888 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.8313010 0.9896272
#> 
#> $U_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.17032236 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01597962 0.07635881 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.17328007 0.2165077 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4206939 0.5173125 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1731996 0.9374756
#> 
#> $U_list[[9]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.43326880 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01528773 0.2057601 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3021501 0.3679075 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2604911 0.5171856 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2263112 0.9787172
#> 
#> $U_list[[10]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.06145321 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.31900255 0.44396571 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.01614824 0.3348477 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1787952 0.2675066 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5052584 0.9120597
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    6   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    9   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    6    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    6   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   10    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    8   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    4    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    9   13   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    7    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(
  fun = "eigs", arg = "lambda", lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07403564 0.00000000 4.0000000 5.0000000 8.0000000
#> [2,] 0.23184165 0.33731042 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08031099 0.4510452 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2360848 0.7332877 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1795544 0.9784429
#> 
#> $A_list[[2]]
#>            [,1]      [,2]       [,3]        [,4]   [,5]
#> [1,] 0.13782087 0.0000000 5.00000000 15.00000000 9.0000
#> [2,] 0.08975026 0.1776651 0.00000000  0.00000000 0.0000
#> [3,] 0.00000000 0.1285642 0.77664094  0.00000000 0.0000
#> [4,] 0.00000000 0.0000000 0.05105833  0.03433595 0.0000
#> [5,] 0.00000000 0.0000000 0.00000000  0.87507406 0.9373
#> 
#> $A_list[[3]]
#>           [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.3251057 0.00000000 3.00000000 8.00000000 8.0000000
#> [2,] 0.1616139 0.51880909 0.00000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.01211708 0.73702339 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.04869819 0.06867132 0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.82783082 0.9447064
#> 
#> $A_list[[4]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0527148 0.0000000 5.0000000 6.0000000 8.0000000
#> [2,] 0.2077534 0.1468495 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2921314 0.2435667 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2416335 0.1137358 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.7593832 0.8898025
#> 
#> $A_list[[5]]
#>             [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.006916893 0.00000000 3.0000000 14.0000000 11.0000000
#> [2,] 0.021080262 0.01532763 0.0000000  0.0000000  0.0000000
#> [3,] 0.000000000 0.03845111 0.1446500  0.0000000  0.0000000
#> [4,] 0.000000000 0.00000000 0.5067203  0.4693682  0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000  0.4624546  0.9670836
#> 
#> $A_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09544632 0.0000000 3.0000000 9.0000000 9.0000000
#> [2,] 0.19895548 0.3340599 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1859213 0.6208345 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1235887 0.4214251 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3237440 0.7731542
#> 
#> $A_list[[7]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1753289 0.0000000 3.0000000 10.0000000 8.0000000
#> [2,] 0.1603206 0.4128156 0.0000000  0.0000000 0.0000000
#> [3,] 0.0000000 0.1044726 0.2026807  0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.5515463  0.5663649 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000  0.2086388 0.8035572
#> 
#> $A_list[[8]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2803295 0.0000000 2.00000000 7.0000000 4.0000000
#> [2,] 0.0759348 0.2848589 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2525334 0.61923021 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.02069761 0.6553533 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3289794 0.9848701
#> 
#> $A_list[[9]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.01765578 0.0000000 5.00000000 8.0000000 8.0000000
#> [2,] 0.14887976 0.1465531 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2293538 0.07551877 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.36799084 0.7571444 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.1484502 0.9592556
#> 
#> $A_list[[10]]
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.2341279 0.0000000 3.0000000 10.0000000 10.0000000
#> [2,] 0.1400723 0.3825989 0.0000000  0.0000000  0.0000000
#> [3,] 0.0000000 0.1612485 0.3254258  0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.2252907  0.7223840  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000  0.1342769  0.8748143
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07403564 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.23184165 0.33731042 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08031099 0.4510452 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2360848 0.7332877 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1795544 0.9784429
#> 
#> $U_list[[2]]
#>            [,1]      [,2]       [,3]       [,4]   [,5]
#> [1,] 0.13782087 0.0000000 0.00000000 0.00000000 0.0000
#> [2,] 0.08975026 0.1776651 0.00000000 0.00000000 0.0000
#> [3,] 0.00000000 0.1285642 0.77664094 0.00000000 0.0000
#> [4,] 0.00000000 0.0000000 0.05105833 0.03433595 0.0000
#> [5,] 0.00000000 0.0000000 0.00000000 0.87507406 0.9373
#> 
#> $U_list[[3]]
#>           [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.3251057 0.00000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.1616139 0.51880909 0.00000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.01211708 0.73702339 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.04869819 0.06867132 0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.82783082 0.9447064
#> 
#> $U_list[[4]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0527148 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2077534 0.1468495 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2921314 0.2435667 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2416335 0.1137358 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.7593832 0.8898025
#> 
#> $U_list[[5]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.006916893 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.021080262 0.01532763 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.03845111 0.1446500 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.5067203 0.4693682 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.4624546 0.9670836
#> 
#> $U_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09544632 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.19895548 0.3340599 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1859213 0.6208345 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1235887 0.4214251 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3237440 0.7731542
#> 
#> $U_list[[7]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1753289 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1603206 0.4128156 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1044726 0.2026807 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.5515463 0.5663649 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2086388 0.8035572
#> 
#> $U_list[[8]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2803295 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.0759348 0.2848589 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2525334 0.61923021 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.02069761 0.6553533 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3289794 0.9848701
#> 
#> $U_list[[9]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.01765578 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.14887976 0.1465531 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2293538 0.07551877 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.36799084 0.7571444 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.1484502 0.9592556
#> 
#> $U_list[[10]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2341279 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1400723 0.3825989 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1612485 0.3254258 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2252907 0.7223840 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.1342769 0.8748143
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    5    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5   15    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   14   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7    4
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08917922 0.000000000 6.0000000 8.0000000 4.0000000
#> [2,] 0.03703057 0.006295257 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.282012257 0.1371438 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.6259290 0.3715448 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.5173176 0.9793781
#> 
#> $A_list[[2]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.36788970 0.0000000 2.0000000 15.0000000 6.0000000
#> [2,] 0.02651439 0.2062995 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.3533624 0.5118801  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3474111  0.5483819 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.3573901 0.9682694
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.06658034 0.00000000 3.0000000 5.00000000 8.0000000
#> [2,] 0.26008019 0.32532872 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.07718801 0.2969669 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4014218 0.88869725 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.04441333 0.9508738
#> 
#> $A_list[[4]]
#>           [,1]     [,2]       [,3]       [,4]      [,5]
#> [1,] 0.3965211 0.000000 2.00000000 9.00000000 16.000000
#> [2,] 0.1083256 0.405207 0.00000000 0.00000000  0.000000
#> [3,] 0.0000000 0.273285 0.04739031 0.00000000  0.000000
#> [4,] 0.0000000 0.000000 0.64861944 0.02775972  0.000000
#> [5,] 0.0000000 0.000000 0.00000000 0.67758590  0.949276
#> 
#> $A_list[[5]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.05261089 0.0000000 2.00000000 9.0000000 7.0000000
#> [2,] 0.39597619 0.4403708 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2012119 0.03521111 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.63750354 0.3693662 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3609503 0.8163445
#> 
#> $A_list[[6]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.06021184 0.00000000 2.0000000 7.0000000 5.0000000
#> [2,] 0.14079044 0.67037622 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.04060909 0.5604017 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2137122 0.3209025 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6061155 0.9726754
#> 
#> $A_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.02356691 0.0000000 3.0000000 9.0000000 14.000000
#> [2,] 0.18804888 0.2521876 0.0000000 0.0000000  0.000000
#> [3,] 0.00000000 0.1700150 0.5033772 0.0000000  0.000000
#> [4,] 0.00000000 0.0000000 0.1045331 0.4581008  0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1784059  0.721701
#> 
#> $A_list[[8]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.23234977 0.0000000 3.00000000 6.0000000 11.0000000
#> [2,] 0.04142177 0.1495680 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2632283 0.72804264 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.02669218 0.1920560  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5923886  0.9823319
#> 
#> $A_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.08547107 0.00000000 6.0000000 5.0000000 15.0000000
#> [2,] 0.07785945 0.04141021 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.57090202 0.4048855 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.2125967 0.2301532  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5881217  0.8272908
#> 
#> $A_list[[10]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.35614618 0.000000000 5.0000000 7.0000000 9.0000000
#> [2,] 0.01352044 0.581453940 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.009145076 0.2731718 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.3930880 0.2309476 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.4456403 0.9467732
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08917922 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03703057 0.006295257 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.282012257 0.1371438 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.6259290 0.3715448 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.5173176 0.9793781
#> 
#> $U_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.36788970 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02651439 0.2062995 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3533624 0.5118801 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3474111 0.5483819 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3573901 0.9682694
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.06658034 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.26008019 0.32532872 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.07718801 0.2969669 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4014218 0.88869725 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.04441333 0.9508738
#> 
#> $U_list[[4]]
#>           [,1]     [,2]       [,3]       [,4]     [,5]
#> [1,] 0.3965211 0.000000 0.00000000 0.00000000 0.000000
#> [2,] 0.1083256 0.405207 0.00000000 0.00000000 0.000000
#> [3,] 0.0000000 0.273285 0.04739031 0.00000000 0.000000
#> [4,] 0.0000000 0.000000 0.64861944 0.02775972 0.000000
#> [5,] 0.0000000 0.000000 0.00000000 0.67758590 0.949276
#> 
#> $U_list[[5]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.05261089 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.39597619 0.4403708 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2012119 0.03521111 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.63750354 0.3693662 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3609503 0.8163445
#> 
#> $U_list[[6]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.06021184 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.14079044 0.67037622 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.04060909 0.5604017 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2137122 0.3209025 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6061155 0.9726754
#> 
#> $U_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]     [,5]
#> [1,] 0.02356691 0.0000000 0.0000000 0.0000000 0.000000
#> [2,] 0.18804888 0.2521876 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.1700150 0.5033772 0.0000000 0.000000
#> [4,] 0.00000000 0.0000000 0.1045331 0.4581008 0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.1784059 0.721701
#> 
#> $U_list[[8]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.23234977 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.04142177 0.1495680 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2632283 0.72804264 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.02669218 0.1920560 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5923886 0.9823319
#> 
#> $U_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08547107 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.07785945 0.04141021 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.57090202 0.4048855 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2125967 0.2301532 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5881217 0.8272908
#> 
#> $U_list[[10]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.35614618 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01352044 0.581453940 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.009145076 0.2731718 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.3930880 0.2309476 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.4456403 0.9467732
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    8    4
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   15    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    5    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    9   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    9    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    6   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    5   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    7    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>