Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfil the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE. Defaults to TRUE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

An integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE, by_type = FALSE
)
#> [[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.39413786 0.00000000 2.0000000 7.0000000 9.0000000
#> [2,] 0.02148542 0.04712157 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.54306076 0.4518025 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2256430 0.1932323 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5766855 0.8508964
#> 
#> [[2]]
#>           [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1618696 0.00000000 5.00000000 4.0000000 12.0000000
#> [2,] 0.1031598 0.54410039 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.06157977 0.57172304 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.04589874 0.6851412  0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.0396190  0.8218046
#> 
#> [[3]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1341243 0.0000000 4.00000000 5.0000000 7.0000000
#> [2,] 0.4054008 0.3537381 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2588298 0.69316426 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.04816166 0.3355461 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.4707688 0.8274001
#> 
#> [[4]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2754621 0.0000000 2.00000000 2.0000000 10.0000000
#> [2,] 0.0785307 0.2449556 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1827455 0.34445395 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.09360598 0.1447526  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.7384928  0.9346462
#> 
#> [[5]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.13248722 0.00000000 2.00000000 5.0000000 8.0000000
#> [2,] 0.05103225 0.08949641 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.48569938 0.68502441 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.07457235 0.5545678 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.3451271 0.9384575
#> 
#> [[6]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.04218191 0.00000000 5.0000000 8.0000000 15.0000000
#> [2,] 0.63543282 0.64336354 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.07366147 0.4979387 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.3002749 0.7668051  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2273968  0.9957362
#> 
#> [[7]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1869641 0.0000000 4.00000000 8.0000000 3.0000000
#> [2,] 0.1741433 0.2641907 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1168374 0.54331034 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.03457146 0.7157580 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2440394 0.9859982
#> 
#> [[8]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1002505 0.0000000 3.0000000 8.0000000 6.0000000
#> [2,] 0.1715855 0.3845287 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1528141 0.3096275 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3070806 0.3599722 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4478171 0.8152792
#> 
#> [[9]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2374239 0.0000000 3.0000000 8.0000000 13.0000000
#> [2,] 0.2554728 0.4339411 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1862464 0.5945231 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.2657374 0.1651876  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.8223851  0.9954345
#> 
#> [[10]]
#>             [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.001813775 0.00000000 1.0000000 7.00000000 8.0000000
#> [2,] 0.134033755 0.08035354 0.0000000 0.00000000 0.0000000
#> [3,] 0.000000000 0.71684706 0.2355919 0.00000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.7076876 0.08296916 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.89770150 0.9834008
#> 

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(
  fun = "lambda", arg = NA, lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.002453827 0.00000000 4.0000000 9.0000000 6.0000000
#> [2,] 0.258116840 0.47813365 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.03635072 0.1066634 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.5445952 0.3119559 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.4584238 0.9445414
#> 
#> $A_list[[2]]
#>            [,1]      [,2]      [,3]        [,4]     [,5]
#> [1,] 0.62974475 0.0000000 8.0000000 6.000000000 13.00000
#> [2,] 0.02272177 0.5825253 0.0000000 0.000000000  0.00000
#> [3,] 0.00000000 0.1311891 0.5030093 0.000000000  0.00000
#> [4,] 0.00000000 0.0000000 0.2239840 0.001300892  0.00000
#> [5,] 0.00000000 0.0000000 0.0000000 0.753659984  0.97989
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.12110590 0.0000000 2.0000000 10.0000000 8.0000000
#> [2,] 0.09151743 0.1368095 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.4140814 0.6604604  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.0267541  0.1438993 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.7718515 0.9858615
#> 
#> $A_list[[4]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.13245563 0.0000000 4.0000000 16.0000000 11.0000000
#> [2,] 0.04397079 0.5540017 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.1505338 0.4779305  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3704536  0.1625879  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.7816190  0.9557546
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]        [,4]       [,5]
#> [1,] 0.18588343 0.00000000 5.0000000 13.00000000 12.0000000
#> [2,] 0.08640561 0.36509551 0.0000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.07809237 0.3444179  0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.3760766  0.81205872  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.03302351  0.8828128
#> 
#> $A_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05530499 0.0000000 6.0000000 4.0000000 10.000000
#> [2,] 0.05202543 0.2515751 0.0000000 0.0000000  0.000000
#> [3,] 0.00000000 0.0572042 0.1534503 0.0000000  0.000000
#> [4,] 0.00000000 0.0000000 0.5175144 0.1536390  0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7323311  0.947502
#> 
#> $A_list[[7]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1658894 0.0000000 2.00000000 9.0000000 9.0000000
#> [2,] 0.1292069 0.2886081 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2414145 0.48105967 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.07044906 0.1031216 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.7118009 0.9522386
#> 
#> $A_list[[8]]
#>             [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.001427132 0.0000000 7.0000000 6.00000000 10.0000000
#> [2,] 0.108819528 0.0417949 0.0000000 0.00000000  0.0000000
#> [3,] 0.000000000 0.4773230 0.4620547 0.00000000  0.0000000
#> [4,] 0.000000000 0.0000000 0.2409149 0.66220908  0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.07429877  0.9607456
#> 
#> $A_list[[9]]
#>           [,1]      [,2]      [,3]        [,4]      [,5]
#> [1,] 0.1172924 0.0000000 3.0000000 10.00000000 6.0000000
#> [2,] 0.0583134 0.2486084 0.0000000  0.00000000 0.0000000
#> [3,] 0.0000000 0.1324760 0.1899529  0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3003975  0.80653096 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000  0.03761629 0.8835781
#> 
#> $A_list[[10]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.11381695 0.00000000 2.00000000 7.0000000 9.0000000
#> [2,] 0.07468044 0.38444342 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08847666 0.09407706 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.77888183 0.4471884 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.5256575 0.9868583
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.002453827 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.258116840 0.47813365 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.03635072 0.1066634 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.5445952 0.3119559 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.4584238 0.9445414
#> 
#> $U_list[[2]]
#>            [,1]      [,2]      [,3]        [,4]    [,5]
#> [1,] 0.62974475 0.0000000 0.0000000 0.000000000 0.00000
#> [2,] 0.02272177 0.5825253 0.0000000 0.000000000 0.00000
#> [3,] 0.00000000 0.1311891 0.5030093 0.000000000 0.00000
#> [4,] 0.00000000 0.0000000 0.2239840 0.001300892 0.00000
#> [5,] 0.00000000 0.0000000 0.0000000 0.753659984 0.97989
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.12110590 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.09151743 0.1368095 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.4140814 0.6604604 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.0267541 0.1438993 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7718515 0.9858615
#> 
#> $U_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.13245563 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04397079 0.5540017 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1505338 0.4779305 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3704536 0.1625879 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7816190 0.9557546
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.18588343 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.08640561 0.36509551 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.07809237 0.3444179 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3760766 0.81205872 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.03302351 0.8828128
#> 
#> $U_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]     [,5]
#> [1,] 0.05530499 0.0000000 0.0000000 0.0000000 0.000000
#> [2,] 0.05202543 0.2515751 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.0572042 0.1534503 0.0000000 0.000000
#> [4,] 0.00000000 0.0000000 0.5175144 0.1536390 0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7323311 0.947502
#> 
#> $U_list[[7]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1658894 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1292069 0.2886081 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2414145 0.48105967 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.07044906 0.1031216 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.7118009 0.9522386
#> 
#> $U_list[[8]]
#>             [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.001427132 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.108819528 0.0417949 0.0000000 0.00000000 0.0000000
#> [3,] 0.000000000 0.4773230 0.4620547 0.00000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.2409149 0.66220908 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.07429877 0.9607456
#> 
#> $U_list[[9]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1172924 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.0583134 0.2486084 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.1324760 0.1899529 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3003975 0.80653096 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.03761629 0.8835781
#> 
#> $U_list[[10]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.11381695 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.07468044 0.38444342 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08847666 0.09407706 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.77888183 0.4471884 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.5256575 0.9868583
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    9    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8    6   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   10    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   16   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5   13   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    4   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    9    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    6   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(
  fun = "eigs", arg = "lambda", lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.28120621 0.00000000 7.0000000 11.0000000 9.0000000
#> [2,] 0.02069694 0.25047513 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.08860227 0.1448429  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2549618  0.5405770 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.4126414 0.9674093
#> 
#> $A_list[[2]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.27366222 0.00000000 5.0000000 8.0000000 8.0000000
#> [2,] 0.04054827 0.08321134 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.48223453 0.5411496 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1765716 0.2087833 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6577939 0.9833212
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01371191 0.00000000 2.0000000 6.0000000 11.000000
#> [2,] 0.07114284 0.15055186 0.0000000 0.0000000  0.000000
#> [3,] 0.00000000 0.05958326 0.0871958 0.0000000  0.000000
#> [4,] 0.00000000 0.00000000 0.4524905 0.4435442  0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.3911497  0.866679
#> 
#> $A_list[[4]]
#>             [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.088127421 0.0000000 6.00000000 14.0000000 7.0000000
#> [2,] 0.006944572 0.1331258 0.00000000  0.0000000 0.0000000
#> [3,] 0.000000000 0.4626101 0.01336594  0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.66942909  0.6167761 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000  0.2453976 0.9301899
#> 
#> $A_list[[5]]
#>             [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.068247862 0.0000000 7.0000000 7.00000000 16.0000000
#> [2,] 0.002068006 0.5022461 0.0000000 0.00000000  0.0000000
#> [3,] 0.000000000 0.2579912 0.6801421 0.00000000  0.0000000
#> [4,] 0.000000000 0.0000000 0.1420713 0.06603674  0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.82971759  0.8999529
#> 
#> $A_list[[6]]
#>             [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.044170993 0.00000000 8.00000000 6.0000000 12.0000000
#> [2,] 0.002607606 0.01646351 0.00000000 0.0000000  0.0000000
#> [3,] 0.000000000 0.35717322 0.04189633 0.0000000  0.0000000
#> [4,] 0.000000000 0.00000000 0.58236869 0.6320868  0.0000000
#> [5,] 0.000000000 0.00000000 0.00000000 0.2431634  0.9446107
#> 
#> $A_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.23940330 0.00000000 1.0000000 6.0000000 9.0000000
#> [2,] 0.01786059 0.47841675 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08302076 0.4047267 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4305993 0.7566412 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1168893 0.9270771
#> 
#> $A_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.01531563 0.00000000 6.0000000 6.0000000 9.000000
#> [2,] 0.09069036 0.08282896 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.47832807 0.2824337 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.4330850 0.4273441 0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.3444451 0.897546
#> 
#> $A_list[[9]]
#>             [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.013997325 0.00000000 5.0000000 6.0000000 11.0000000
#> [2,] 0.009202573 0.06829169 0.0000000 0.0000000  0.0000000
#> [3,] 0.000000000 0.30154345 0.1239476 0.0000000  0.0000000
#> [4,] 0.000000000 0.00000000 0.5508953 0.2281621  0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.6120016  0.8781885
#> 
#> $A_list[[10]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.31253229 0.0000000 5.00000000 9.0000000 13.0000000
#> [2,] 0.09645687 0.1666158 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3368785 0.72675039 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.04473588 0.4866409  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3570152  0.9677079
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.28120621 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02069694 0.25047513 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08860227 0.1448429 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2549618 0.5405770 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.4126414 0.9674093
#> 
#> $U_list[[2]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.27366222 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04054827 0.08321134 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.48223453 0.5411496 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1765716 0.2087833 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.6577939 0.9833212
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.01371191 0.00000000 0.0000000 0.0000000 0.000000
#> [2,] 0.07114284 0.15055186 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.05958326 0.0871958 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.4524905 0.4435442 0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.3911497 0.866679
#> 
#> $U_list[[4]]
#>             [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.088127421 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.006944572 0.1331258 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.4626101 0.01336594 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.66942909 0.6167761 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.2453976 0.9301899
#> 
#> $U_list[[5]]
#>             [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.068247862 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.002068006 0.5022461 0.0000000 0.00000000 0.0000000
#> [3,] 0.000000000 0.2579912 0.6801421 0.00000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.1420713 0.06603674 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.82971759 0.8999529
#> 
#> $U_list[[6]]
#>             [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.044170993 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.002607606 0.01646351 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.35717322 0.04189633 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.58236869 0.6320868 0.0000000
#> [5,] 0.000000000 0.00000000 0.00000000 0.2431634 0.9446107
#> 
#> $U_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.23940330 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.01786059 0.47841675 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08302076 0.4047267 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4305993 0.7566412 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1168893 0.9270771
#> 
#> $U_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.01531563 0.00000000 0.0000000 0.0000000 0.000000
#> [2,] 0.09069036 0.08282896 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.47832807 0.2824337 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.4330850 0.4273441 0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.3444451 0.897546
#> 
#> $U_list[[9]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.013997325 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.009202573 0.06829169 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.30154345 0.1239476 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.5508953 0.2281621 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.6120016 0.8781885
#> 
#> $U_list[[10]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.31253229 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.09645687 0.1666158 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3368785 0.72675039 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.04473588 0.4866409 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3570152 0.9677079
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7   11    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   14    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    7   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8    6   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    6    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    6    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    9   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1704326 0.0000000 6.0000000 5.0000000 11.0000000
#> [2,] 0.2526541 0.1788381 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.5562997 0.5179539 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3440421 0.7417104  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.1349894  0.8814724
#> 
#> $A_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.33632032 0.0000000 2.0000000 7.0000000 13.0000000
#> [2,] 0.08321032 0.4118319 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1014435 0.4212170 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2151534 0.4888852  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3021386  0.9255292
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.08893095 0.0000000 1.0000000 5.0000000 15.0000000
#> [2,] 0.04844071 0.6112252 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2650061 0.6597279 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2599914 0.5947452  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3665834  0.9700162
#> 
#> $A_list[[4]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.01639093 0.0000000 2.0000000 10.0000000 12.0000000
#> [2,] 0.24721497 0.1235475 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.2873161 0.5780632  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.2246370  0.5391206  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.3755137  0.9168869
#> 
#> $A_list[[5]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.4416876 0.0000000 7.00000000 6.0000000 7.0000000
#> [2,] 0.3002683 0.1842014 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.5651136 0.73565235 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.02296356 0.5749099 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3154161 0.9139486
#> 
#> $A_list[[6]]
#>             [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.004046487 0.0000000 5.00000000 7.00000000 3.0000000
#> [2,] 0.012151932 0.1104949 0.00000000 0.00000000 0.0000000
#> [3,] 0.000000000 0.1119789 0.56300119 0.00000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.01693818 0.07768281 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.52471635 0.8969879
#> 
#> $A_list[[7]]
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.22478538 0.00000000 8.0000000 12.0000000 11.0000000
#> [2,] 0.07713295 0.38828871 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.07148182 0.3938964  0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.3833508  0.6773859  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.1980568  0.8944151
#> 
#> $A_list[[8]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.4824876 0.0000000 2.00000000 5.0000000 8.0000000
#> [2,] 0.1199741 0.6576708 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.0603594 0.05202994 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.70847109 0.6257849 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.1748572 0.8228055
#> 
#> $A_list[[9]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.06868318 0.00000000 3.0000000 13.0000000 11.000000
#> [2,] 0.14637691 0.08787371 0.0000000  0.0000000  0.000000
#> [3,] 0.00000000 0.57839642 0.3650416  0.0000000  0.000000
#> [4,] 0.00000000 0.00000000 0.3714341  0.7128707  0.000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.1697840  0.900483
#> 
#> $A_list[[10]]
#>            [,1]        [,2]      [,3]        [,4]      [,5]
#> [1,] 0.61710105 0.000000000 3.0000000 5.000000000 7.0000000
#> [2,] 0.05358025 0.677134119 0.0000000 0.000000000 0.0000000
#> [3,] 0.00000000 0.003420948 0.5372546 0.000000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.3555826 0.008842287 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.951347845 0.9657712
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1704326 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2526541 0.1788381 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.5562997 0.5179539 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3440421 0.7417104 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.1349894 0.8814724
#> 
#> $U_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.33632032 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.08321032 0.4118319 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1014435 0.4212170 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2151534 0.4888852 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3021386 0.9255292
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08893095 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04844071 0.6112252 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2650061 0.6597279 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2599914 0.5947452 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3665834 0.9700162
#> 
#> $U_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01639093 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.24721497 0.1235475 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2873161 0.5780632 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.2246370 0.5391206 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3755137 0.9168869
#> 
#> $U_list[[5]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.4416876 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.3002683 0.1842014 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.5651136 0.73565235 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.02296356 0.5749099 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3154161 0.9139486
#> 
#> $U_list[[6]]
#>             [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.004046487 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.012151932 0.1104949 0.00000000 0.00000000 0.0000000
#> [3,] 0.000000000 0.1119789 0.56300119 0.00000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.01693818 0.07768281 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.52471635 0.8969879
#> 
#> $U_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.22478538 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.07713295 0.38828871 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.07148182 0.3938964 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3833508 0.6773859 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1980568 0.8944151
#> 
#> $U_list[[8]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.4824876 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1199741 0.6576708 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.0603594 0.05202994 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.70847109 0.6257849 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.1748572 0.8228055
#> 
#> $U_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.06868318 0.00000000 0.0000000 0.0000000 0.000000
#> [2,] 0.14637691 0.08787371 0.0000000 0.0000000 0.000000
#> [3,] 0.00000000 0.57839642 0.3650416 0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.3714341 0.7128707 0.000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1697840 0.900483
#> 
#> $U_list[[10]]
#>            [,1]        [,2]      [,3]        [,4]      [,5]
#> [1,] 0.61710105 0.000000000 0.0000000 0.000000000 0.0000000
#> [2,] 0.05358025 0.677134119 0.0000000 0.000000000 0.0000000
#> [3,] 0.00000000 0.003420948 0.5372546 0.000000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.3555826 0.008842287 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.951347845 0.9657712
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    5   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    5   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   10   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    6    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    7    3
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8   12   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    5    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   13   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    5    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>