Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfil the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE. Defaults to TRUE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

An integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE, by_type = FALSE
)
#> [[1]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1388397 0.0000000 2.00000000 6.0000000 9.0000000
#> [2,] 0.6111329 0.4220730 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3674604 0.83312035 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.01868955 0.5221617 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.4677711 0.9945742
#> 
#> [[2]]
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1717924 0.0000000 3.0000000 14.0000000 12.0000000
#> [2,] 0.3533404 0.1205226 0.0000000  0.0000000  0.0000000
#> [3,] 0.0000000 0.5026770 0.2916456  0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.4542227  0.3938196  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000  0.5202995  0.9826523
#> 
#> [[3]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.04107501 0.0000000 1.0000000 4.0000000 7.0000000
#> [2,] 0.19618726 0.1299740 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1404766 0.3063208 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3278050 0.1791563 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6614806 0.8705798
#> 
#> [[4]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.68005620 0.0000000 2.00000000 4.0000000 12.0000000
#> [2,] 0.02092037 0.4933634 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2804563 0.76314241 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.03144515 0.5881464  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3242738  0.9546347
#> 
#> [[5]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.08101023 0.0000000 2.0000000 12.0000000 10.0000000
#> [2,] 0.09896621 0.2476765 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.2198743 0.2101915  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.5426414  0.7510175  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.2237568  0.9907616
#> 
#> [[6]]
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.06941807 0.00000000 5.00000000 6.0000000 10.0000000
#> [2,] 0.16553111 0.02145593 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.46013821 0.54315157 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.06897589 0.6519617  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.3091108  0.9993208
#> 
#> [[7]]
#>           [,1]       [,2]      [,3]        [,4]      [,5]
#> [1,] 0.2289883 0.00000000 1.0000000 15.00000000 7.0000000
#> [2,] 0.0564396 0.22177105 0.0000000  0.00000000 0.0000000
#> [3,] 0.0000000 0.09092593 0.0179237  0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.5719557  0.89680216 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000  0.07292402 0.9902836
#> 
#> [[8]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.19240661 0.0000000 3.00000000 9.00000000 8.0000000
#> [2,] 0.07715826 0.1959947 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2957537 0.62770378 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.07757236 0.09795369 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.77318143 0.9104562
#> 
#> [[9]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.04656685 0.0000000 4.0000000 7.0000000 16.0000000
#> [2,] 0.74247539 0.4580732 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.4182243 0.5587630 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3453089 0.3349983  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5946952  0.9379546
#> 
#> [[10]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.2585374 0.0000000 5.0000000 10.0000000 5.0000000
#> [2,] 0.3440145 0.4463790 0.0000000  0.0000000 0.0000000
#> [3,] 0.0000000 0.2094757 0.2939255  0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.5833267  0.8198305 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000  0.1295853 0.9585671
#> 

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(
  fun = "lambda", arg = NA, lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.34721331 0.0000000 6.00000000 9.0000000 6.0000000
#> [2,] 0.09611996 0.5024694 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1220555 0.06400727 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.64437806 0.3856262 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5052933 0.9151682
#> 
#> $A_list[[2]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.4625605 0.0000000 2.0000000 7.0000000 15.000000
#> [2,] 0.0203033 0.1962159 0.0000000 0.0000000  0.000000
#> [3,] 0.0000000 0.3592674 0.3506185 0.0000000  0.000000
#> [4,] 0.0000000 0.0000000 0.2562135 0.4313121  0.000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4296372  0.938268
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.01708142 0.00000000 2.0000000 11.0000000 6.0000000
#> [2,] 0.14328512 0.49934464 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.02041444 0.5601248  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3005189  0.6401858 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.2518518 0.9479214
#> 
#> $A_list[[4]]
#>           [,1]        [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1040288 0.000000000 8.00000000 6.0000000 10.0000000
#> [2,] 0.1494574 0.485799471 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.009031298 0.08009629 0.0000000  0.0000000
#> [4,] 0.0000000 0.000000000 0.73160328 0.3280418  0.0000000
#> [5,] 0.0000000 0.000000000 0.00000000 0.5029728  0.9138576
#> 
#> $A_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.25883024 0.0000000 3.0000000 9.0000000 11.0000000
#> [2,] 0.06905594 0.2966454 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1889333 0.1855742 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.5763560 0.5340340  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2534221  0.9606919
#> 
#> $A_list[[6]]
#>           [,1]        [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2287163 0.000000000 5.0000000 7.0000000 10.0000000
#> [2,] 0.2666265 0.530912636 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.009419193 0.3912374 0.0000000  0.0000000
#> [4,] 0.0000000 0.000000000 0.1568816 0.7085584  0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.1400723  0.8761656
#> 
#> $A_list[[7]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1708085 0.00000000 2.0000000 10.0000000 11.000000
#> [2,] 0.1610661 0.47595679 0.0000000  0.0000000  0.000000
#> [3,] 0.0000000 0.04497979 0.1698503  0.0000000  0.000000
#> [4,] 0.0000000 0.00000000 0.4135316  0.7771748  0.000000
#> [5,] 0.0000000 0.00000000 0.0000000  0.1467766  0.944207
#> 
#> $A_list[[8]]
#>            [,1]       [,2]        [,3]       [,4]      [,5]
#> [1,] 0.11907748 0.00000000 2.000000000 11.0000000 7.0000000
#> [2,] 0.06172065 0.08851621 0.000000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.20297194 0.001314552  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.294578613  0.5033933 0.0000000
#> [5,] 0.00000000 0.00000000 0.000000000  0.3530127 0.9811423
#> 
#> $A_list[[9]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.16653139 0.0000000 3.0000000 3.0000000 7.0000000
#> [2,] 0.00747946 0.5140402 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1258477 0.2119491 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.6693878 0.2983609 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6253412 0.9288982
#> 
#> $A_list[[10]]
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.08956904 0.00000000 5.00000000 9.0000000 11.0000000
#> [2,] 0.12284224 0.59884824 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.03079704 0.01681033 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.77638545 0.1994359  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6724318  0.9315394
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.34721331 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.09611996 0.5024694 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1220555 0.06400727 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.64437806 0.3856262 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5052933 0.9151682
#> 
#> $U_list[[2]]
#>           [,1]      [,2]      [,3]      [,4]     [,5]
#> [1,] 0.4625605 0.0000000 0.0000000 0.0000000 0.000000
#> [2,] 0.0203033 0.1962159 0.0000000 0.0000000 0.000000
#> [3,] 0.0000000 0.3592674 0.3506185 0.0000000 0.000000
#> [4,] 0.0000000 0.0000000 0.2562135 0.4313121 0.000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4296372 0.938268
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01708142 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.14328512 0.49934464 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02041444 0.5601248 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3005189 0.6401858 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2518518 0.9479214
#> 
#> $U_list[[4]]
#>           [,1]        [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1040288 0.000000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1494574 0.485799471 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.009031298 0.08009629 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.73160328 0.3280418 0.0000000
#> [5,] 0.0000000 0.000000000 0.00000000 0.5029728 0.9138576
#> 
#> $U_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.25883024 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.06905594 0.2966454 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1889333 0.1855742 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.5763560 0.5340340 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2534221 0.9606919
#> 
#> $U_list[[6]]
#>           [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2287163 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2666265 0.530912636 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.009419193 0.3912374 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.1568816 0.7085584 0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.1400723 0.8761656
#> 
#> $U_list[[7]]
#>           [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.1708085 0.00000000 0.0000000 0.0000000 0.000000
#> [2,] 0.1610661 0.47595679 0.0000000 0.0000000 0.000000
#> [3,] 0.0000000 0.04497979 0.1698503 0.0000000 0.000000
#> [4,] 0.0000000 0.00000000 0.4135316 0.7771748 0.000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.1467766 0.944207
#> 
#> $U_list[[8]]
#>            [,1]       [,2]        [,3]      [,4]      [,5]
#> [1,] 0.11907748 0.00000000 0.000000000 0.0000000 0.0000000
#> [2,] 0.06172065 0.08851621 0.000000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.20297194 0.001314552 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.294578613 0.5033933 0.0000000
#> [5,] 0.00000000 0.00000000 0.000000000 0.3530127 0.9811423
#> 
#> $U_list[[9]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.16653139 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.00747946 0.5140402 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1258477 0.2119491 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.6693878 0.2983609 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6253412 0.9288982
#> 
#> $U_list[[10]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.08956904 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.12284224 0.59884824 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03079704 0.01681033 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.77638545 0.1994359 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6724318 0.9315394
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    9    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    7   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   11    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8    6   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    7   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   10   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   11    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    3    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    9   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(
  fun = "eigs", arg = "lambda", lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.28033840 0.0000000 0.0000000 4.0000000 13.0000000
#> [2,] 0.05826547 0.2159153 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1765807 0.2976307 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.4581567 0.6094801  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2358149  0.8931403
#> 
#> $A_list[[2]]
#>           [,1]      [,2]      [,3]        [,4]       [,5]
#> [1,] 0.1102475 0.0000000 5.0000000 6.000000000 15.0000000
#> [2,] 0.2707152 0.3941670 0.0000000 0.000000000  0.0000000
#> [3,] 0.0000000 0.1319533 0.1577155 0.000000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.4377875 0.724220754  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.005405149  0.8581774
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.11120561 0.0000000 4.0000000 5.0000000 12.0000000
#> [2,] 0.05047238 0.1063720 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3402397 0.3055521 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1701990 0.4503377  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3652667  0.9652679
#> 
#> $A_list[[4]]
#>             [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.627466330 0.0000000 4.0000000 5.00000000 13.0000000
#> [2,] 0.006292947 0.4424895 0.0000000 0.00000000  0.0000000
#> [3,] 0.000000000 0.2765170 0.6233693 0.00000000  0.0000000
#> [4,] 0.000000000 0.0000000 0.2134748 0.08302325  0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.81127210  0.8999988
#> 
#> $A_list[[5]]
#>            [,1]      [,2]      [,3]        [,4]       [,5]
#> [1,] 0.18821459 0.0000000 0.0000000 12.00000000 13.0000000
#> [2,] 0.02269693 0.1864562 0.0000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.2102979 0.3691414  0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.4477073  0.77610357  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.08264141  0.9173613
#> 
#> $A_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.35594650 0.0000000 6.0000000 8.0000000 5.0000000
#> [2,] 0.05606303 0.6442860 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1089456 0.6195987 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1784518 0.2628108 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6252934 0.9264035
#> 
#> $A_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.10412345 0.0000000 2.0000000 6.0000000 11.0000000
#> [2,] 0.04897732 0.1372727 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2337532 0.2243216 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.4644966 0.5368809  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3514062  0.9159737
#> 
#> $A_list[[8]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.23606683 0.0000000 3.0000000 5.0000000 10.0000000
#> [2,] 0.02472292 0.5772939 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2489640 0.1222832 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.7672877 0.2414671  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6698613  0.9746301
#> 
#> $A_list[[9]]
#>            [,1]      [,2]      [,3]        [,4]      [,5]
#> [1,] 0.32155222 0.0000000 2.0000000 11.00000000 9.0000000
#> [2,] 0.02119937 0.1614808 0.0000000  0.00000000 0.0000000
#> [3,] 0.00000000 0.2224946 0.3313500  0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4346557  0.79216297 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.01716158 0.9529799
#> 
#> $A_list[[10]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.07535596 0.0000000 3.0000000 15.0000000 8.0000000
#> [2,] 0.08111144 0.0491135 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.2199404 0.1636633  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1982936  0.3212850 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.5861087 0.9858666
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.28033840 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.05826547 0.2159153 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1765807 0.2976307 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4581567 0.6094801 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2358149 0.8931403
#> 
#> $U_list[[2]]
#>           [,1]      [,2]      [,3]        [,4]      [,5]
#> [1,] 0.1102475 0.0000000 0.0000000 0.000000000 0.0000000
#> [2,] 0.2707152 0.3941670 0.0000000 0.000000000 0.0000000
#> [3,] 0.0000000 0.1319533 0.1577155 0.000000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.4377875 0.724220754 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.005405149 0.8581774
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.11120561 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.05047238 0.1063720 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3402397 0.3055521 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1701990 0.4503377 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3652667 0.9652679
#> 
#> $U_list[[4]]
#>             [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.627466330 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.006292947 0.4424895 0.0000000 0.00000000 0.0000000
#> [3,] 0.000000000 0.2765170 0.6233693 0.00000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.2134748 0.08302325 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.81127210 0.8999988
#> 
#> $U_list[[5]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.18821459 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.02269693 0.1864562 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2102979 0.3691414 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4477073 0.77610357 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.08264141 0.9173613
#> 
#> $U_list[[6]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.35594650 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.05606303 0.6442860 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1089456 0.6195987 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1784518 0.2628108 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6252934 0.9264035
#> 
#> $U_list[[7]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.10412345 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04897732 0.1372727 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2337532 0.2243216 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4644966 0.5368809 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3514062 0.9159737
#> 
#> $U_list[[8]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.23606683 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02472292 0.5772939 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2489640 0.1222832 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.7672877 0.2414671 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6698613 0.9746301
#> 
#> $U_list[[9]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.32155222 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.02119937 0.1614808 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2224946 0.3313500 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4346557 0.79216297 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.01716158 0.9529799
#> 
#> $U_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07535596 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.08111144 0.0491135 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2199404 0.1636633 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1982936 0.3212850 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5861087 0.9858666
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    4   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    5   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    5   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0   12   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    8    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    5   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   11    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   15    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.0260410 0.00000000 4.0000000 12.0000000 13.0000000
#> [2,] 0.4881952 0.09494893 0.0000000  0.0000000  0.0000000
#> [3,] 0.0000000 0.50947153 0.5378430  0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.1086399  0.1086846  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000  0.6591700  0.8709216
#> 
#> $A_list[[2]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.3015111 0.0000000 1.00000000 7.0000000 15.0000000
#> [2,] 0.2914713 0.1652251 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.4735754 0.09908765 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.68070749 0.6618568  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2740694  0.9607499
#> 
#> $A_list[[3]]
#>           [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1357165 0.00000000 2.0000000 8.0000000 10.0000000
#> [2,] 0.1093172 0.45456830 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.05780552 0.2567048 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.4148097 0.3134829  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.5350099  0.9705674
#> 
#> $A_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.03369960 0.0000000 6.0000000 9.0000000 14.0000000
#> [2,] 0.02567207 0.1892297 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2532775 0.2090129 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.4916739 0.6292082  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2192542  0.8640882
#> 
#> $A_list[[5]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2243555 0.0000000 4.0000000 6.0000000 13.0000000
#> [2,] 0.4326043 0.4626014 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.2349708 0.3781518 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3915624 0.5728437  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2333051  0.9735823
#> 
#> $A_list[[6]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1264420 0.00000000 3.0000000 6.00000000 11.0000000
#> [2,] 0.2599201 0.48300244 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.03536851 0.4055713 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.2220586 0.64916925  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.03645881  0.8485933
#> 
#> $A_list[[7]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1557862 0.0000000 6.00000000 9.0000000 17.0000000
#> [2,] 0.0474156 0.4744848 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1520986 0.03054312 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.83588977 0.6211433  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2694360  0.9350979
#> 
#> $A_list[[8]]
#>            [,1]      [,2]      [,3]        [,4]       [,5]
#> [1,] 0.07880031 0.0000000 2.0000000 11.00000000 16.0000000
#> [2,] 0.13659552 0.5492367 0.0000000  0.00000000  0.0000000
#> [3,] 0.00000000 0.2423061 0.6780284  0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1384049  0.05016142  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.87136510  0.9540785
#> 
#> $A_list[[9]]
#>            [,1]         [,2]       [,3]      [,4]       [,5]
#> [1,] 0.23771521 0.000000e+00 2.00000000 9.0000000 11.0000000
#> [2,] 0.02893665 6.218268e-01 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 3.049663e-05 0.03075212 0.0000000  0.0000000
#> [4,] 0.00000000 0.000000e+00 0.73960569 0.6429939  0.0000000
#> [5,] 0.00000000 0.000000e+00 0.00000000 0.2361105  0.9307507
#> 
#> $A_list[[10]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1516920 0.0000000 3.0000000 7.0000000 8.0000000
#> [2,] 0.1402917 0.1881856 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1302858 0.2275192 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.4248372 0.3915107 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.3147972 0.9683852
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.0260410 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.4881952 0.09494893 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.50947153 0.5378430 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.1086399 0.1086846 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.6591700 0.8709216
#> 
#> $U_list[[2]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.3015111 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.2914713 0.1652251 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.4735754 0.09908765 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.68070749 0.6618568 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2740694 0.9607499
#> 
#> $U_list[[3]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1357165 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1093172 0.45456830 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.05780552 0.2567048 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.4148097 0.3134829 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.5350099 0.9705674
#> 
#> $U_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.03369960 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.02567207 0.1892297 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2532775 0.2090129 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4916739 0.6292082 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2192542 0.8640882
#> 
#> $U_list[[5]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2243555 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.4326043 0.4626014 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2349708 0.3781518 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3915624 0.5728437 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2333051 0.9735823
#> 
#> $U_list[[6]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1264420 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.2599201 0.48300244 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.03536851 0.4055713 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.2220586 0.64916925 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.03645881 0.8485933
#> 
#> $U_list[[7]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1557862 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.0474156 0.4744848 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1520986 0.03054312 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.83588977 0.6211433 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2694360 0.9350979
#> 
#> $U_list[[8]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.07880031 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.13659552 0.5492367 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2423061 0.6780284 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1384049 0.05016142 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.87136510 0.9540785
#> 
#> $U_list[[9]]
#>            [,1]         [,2]       [,3]      [,4]      [,5]
#> [1,] 0.23771521 0.000000e+00 0.00000000 0.0000000 0.0000000
#> [2,] 0.02893665 6.218268e-01 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 3.049663e-05 0.03075212 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000e+00 0.73960569 0.6429939 0.0000000
#> [5,] 0.00000000 0.000000e+00 0.00000000 0.2361105 0.9307507
#> 
#> $U_list[[10]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1516920 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1402917 0.1881856 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1302858 0.2275192 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.4248372 0.3915107 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.3147972 0.9683852
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   12   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    7   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    9   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    6   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    6   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    9   17
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   11   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    9   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    7    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>