Generate lists of matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable lambda values. The function attempts to find matrices that fulfill the lambda criteria, discarding unacceptable matrices. If it takes more than 1000 attempts to find a suitable matrix model, then an error is produced.

Usage

generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99
)

Arguments

n

The number of MPMs to generate. Default is 10.

lower_lambda

The lower bound for acceptable lambda values. Default is 0.9.

upper_lambda

The upper bound for acceptable lambda values. Default is 1.1.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list

max_surv

The maximum acceptable survival value. Defaults to 0.99.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)
#> $A_list
#> $A_list[[1]]
#>             [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.202340746 0.0000000 5.00000000 4.0000000 5.0000000
#> [2,] 0.003716639 0.1123783 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.3497316 0.49105790 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.05471826 0.2963376 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.3985038 0.9459818
#> 
#> $A_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.19683756 0.0000000 1.0000000 6.0000000 5.0000000
#> [2,] 0.03537017 0.1636587 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1667347 0.5455549 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1984403 0.2269822 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6357139 0.9591515
#> 
#> $A_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09493135 0.00000000 6.0000000 7.0000000 6.0000000
#> [2,] 0.18631523 0.53394191 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03167485 0.0979098 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.5496196 0.6419382 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2728588 0.9401529
#> 
#> $A_list[[4]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.2078810 0.00000000 6.0000000 7.00000000 10.0000000
#> [2,] 0.1279678 0.53292405 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.04986383 0.3512287 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.2942105 0.08954527  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.64764664  0.8010677
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.04187711 0.00000000 1.0000000 8.00000000 7.0000000
#> [2,] 0.15289423 0.29615298 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.09506744 0.1382293 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4009224 0.68254424 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.05680463 0.9409587
#> 
#> $A_list[[6]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.038311777 0.00000000 6.0000000 7.0000000 7.0000000
#> [2,] 0.002951738 0.07814344 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.25059722 0.5534203 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.1692322 0.5413409 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.3990726 0.9427458
#> 
#> $A_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.24568920 0.00000000 3.0000000 9.0000000 9.0000000
#> [2,] 0.03772062 0.01564645 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.34489008 0.2061581 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.6405615 0.3473121 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5806056 0.9384691
#> 
#> $A_list[[8]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.01782931 0.00000000 3.00000000 8.00000000 9.0000000
#> [2,] 0.34368416 0.31908514 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.04579386 0.54406118 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.03650706 0.57869640 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.03839084 0.9413531
#> 
#> $A_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09010613 0.00000000 3.0000000 9.0000000 7.0000000
#> [2,] 0.15872477 0.25040543 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.09891494 0.1447889 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2818452 0.4577810 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2468258 0.9817911
#> 
#> $A_list[[10]]
#>            [,1]         [,2]     [,3]       [,4]      [,5]
#> [1,] 0.02947349 0.0000000000 4.000000 11.0000000 9.0000000
#> [2,] 0.17368081 0.2432731341 0.000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.0008570823 0.217286  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000000 0.200349  0.3360345 0.0000000
#> [5,] 0.00000000 0.0000000000 0.000000  0.6067448 0.9689789
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>             [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.202340746 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.003716639 0.1123783 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.3497316 0.49105790 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.05471826 0.2963376 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.3985038 0.9459818
#> 
#> $U_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.19683756 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03537017 0.1636587 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1667347 0.5455549 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1984403 0.2269822 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.6357139 0.9591515
#> 
#> $U_list[[3]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09493135 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.18631523 0.53394191 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03167485 0.0979098 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.5496196 0.6419382 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2728588 0.9401529
#> 
#> $U_list[[4]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.2078810 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.1279678 0.53292405 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.04986383 0.3512287 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.2942105 0.08954527 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.64764664 0.8010677
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.04187711 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.15289423 0.29615298 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.09506744 0.1382293 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4009224 0.68254424 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.05680463 0.9409587
#> 
#> $U_list[[6]]
#>             [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.038311777 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.002951738 0.07814344 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.25059722 0.5534203 0.0000000 0.0000000
#> [4,] 0.000000000 0.00000000 0.1692322 0.5413409 0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.3990726 0.9427458
#> 
#> $U_list[[7]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.24568920 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03772062 0.01564645 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.34489008 0.2061581 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.6405615 0.3473121 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.5806056 0.9384691
#> 
#> $U_list[[8]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.01782931 0.00000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.34368416 0.31908514 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.04579386 0.54406118 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.03650706 0.57869640 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.03839084 0.9413531
#> 
#> $U_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09010613 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.15872477 0.25040543 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.09891494 0.1447889 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2818452 0.4577810 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2468258 0.9817911
#> 
#> $U_list[[10]]
#>            [,1]         [,2]     [,3]      [,4]      [,5]
#> [1,] 0.02947349 0.0000000000 0.000000 0.0000000 0.0000000
#> [2,] 0.17368081 0.2432731341 0.000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.0008570823 0.217286 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000000 0.200349 0.3360345 0.0000000
#> [5,] 0.00000000 0.0000000000 0.000000 0.6067448 0.9689789
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    4    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    6    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    7    6
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    7   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    8    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    7    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    8    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    9    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   11    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>