Generate lists of matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable lambda values. The function attempts to find matrices that fulfill the lambda criteria, discarding unacceptable matrices. If it takes more than 1000 attempts to find a suitable matrix model, then an error is produced.

Usage

generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = FALSE
)

Arguments

n

The number of MPMs to generate. Default is 10.

lower_lambda

The lower bound for acceptable lambda values. Default is 0.9.

upper_lambda

The upper bound for acceptable lambda values. Default is 1.1.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split matrices. Default is FALSE.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)
#> [[1]]
#> [[1]]$mat_U
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1238979 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1633336 0.3587661 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1851652 0.06301411 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.48848150 0.1979950 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.4231736 0.7844934
#> 
#> [[1]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    5   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[2]]
#> [[2]]$mat_U
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07266122 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.09772911 0.2428625 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1139768 0.3343656 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1586776 0.5356105 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3359341 0.8968079
#> 
#> [[2]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    9   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[3]]
#> [[3]]$mat_U
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.03144471 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03851080 0.04651266 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.66670453 0.3947482 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3374254 0.5698094 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.3358911 0.9538923
#> 
#> [[3]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[4]]
#> [[4]]$mat_U
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.14496944 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.06254335 0.75198305 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.03245779 0.1003747 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.7412024 0.6814011 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.2671005 0.9863134
#> 
#> [[4]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    4   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[5]]
#> [[5]]$mat_U
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.14709452 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.03929466 0.5094423 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1224438 0.78937178 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.01688883 0.2922895 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5457573 0.9857245
#> 
#> [[5]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   14   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[6]]
#> [[6]]$mat_U
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1496640 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2138491 0.3137306 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2255388 0.2863925 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3519575 0.4548650 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.1979043 0.7687896
#> 
#> [[6]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    4    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[7]]
#> [[7]]$mat_U
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.3448706 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.0149733 0.4886427 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2655843 0.7190523 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.0399443 0.5198091 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4087604 0.9689907
#> 
#> [[7]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    8   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[8]]
#> [[8]]$mat_U
#>             [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.074119127 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.004978316 0.1201826 0.00000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.4804382 0.66484232 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.01762799 0.3117515 0.0000000
#> [5,] 0.000000000 0.0000000 0.00000000 0.6052336 0.9392129
#> 
#> [[8]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    4    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[9]]
#> [[9]]$mat_U
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.26117745 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.01368854 0.2897971 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3658991 0.06364465 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.67396542 0.6970729 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.1546597 0.9421598
#> 
#> [[9]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5   11    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 
#> [[10]]
#> [[10]]$mat_U
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08820801 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.06444837 0.28869927 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.08338718 0.3411485 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3685804 0.7386661 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1632794 0.9688617
#> 
#> [[10]]$mat_F
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8   11   12
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>