Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfil the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must also be FALSE.

max_surv

The maximum acceptable survival value. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

An integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE, by_type = FALSE
)
#> [[1]]
#>             [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.006653799 0.00000000 4.00000000 7.0000000 13.000000
#> [2,] 0.140997173 0.07426255 0.00000000 0.0000000  0.000000
#> [3,] 0.000000000 0.18284467 0.55373279 0.0000000  0.000000
#> [4,] 0.000000000 0.00000000 0.01521576 0.5592577  0.000000
#> [5,] 0.000000000 0.00000000 0.00000000 0.3034959  0.906304
#> 
#> [[2]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.4406195 0.0000000 8.0000000 6.0000000 7.0000000
#> [2,] 0.1015972 0.2577722 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3548119 0.1743998 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.5080130 0.6436063 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.3028826 0.9675714
#> 
#> [[3]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2686022 0.0000000 1.00000000 6.0000000 12.0000000
#> [2,] 0.2317069 0.1801207 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.4469170 0.03306451 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.60196628 0.4617393  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3096062  0.8412271
#> 
#> [[4]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.08736087 0.0000000 1.0000000 12.0000000 9.0000000
#> [2,] 0.29598751 0.1072529 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.5678899 0.2896367  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4769367  0.6962921 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.1815516 0.9359678
#> 
#> [[5]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.09893344 0.0000000 6.0000000 6.0000000 12.0000000
#> [2,] 0.16308602 0.3089759 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1172321 0.5593113 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.0488913 0.2850819  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5779291  0.9417634
#> 
#> [[6]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05345318 0.00000000 3.0000000 9.0000000 7.0000000
#> [2,] 0.18203412 0.39603122 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.06072724 0.3185990 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2178828 0.1741796 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.3647599 0.8126812
#> 
#> [[7]]
#>            [,1]       [,2]      [,3]       [,4]     [,5]
#> [1,] 0.09616868 0.00000000 6.0000000 11.0000000 8.000000
#> [2,] 0.06970082 0.32380588 0.0000000  0.0000000 0.000000
#> [3,] 0.00000000 0.04246682 0.2104377  0.0000000 0.000000
#> [4,] 0.00000000 0.00000000 0.2824999  0.5908180 0.000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.2613032 0.855635
#> 
#> [[8]]
#>           [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1483351 0.00000000 4.00000000 6.0000000 10.0000000
#> [2,] 0.1098040 0.02204395 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.24082812 0.52078986 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.05957025 0.1200780  0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.5342993  0.8418172
#> 
#> [[9]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.08904452 0.0000000 2.00000000 5.0000000 13.000000
#> [2,] 0.01240675 0.2159361 0.00000000 0.0000000  0.000000
#> [3,] 0.00000000 0.4817877 0.74652068 0.0000000  0.000000
#> [4,] 0.00000000 0.0000000 0.06475342 0.2109492  0.000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.7143539  0.936546
#> 
#> [[10]]
#>           [,1]      [,2]        [,3]       [,4]      [,5]
#> [1,] 0.4048289 0.0000000 5.000000000 7.00000000 7.0000000
#> [2,] 0.1470779 0.4202720 0.000000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.2895199 0.750625952 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.003923884 0.04747595 0.0000000
#> [5,] 0.0000000 0.0000000 0.000000000 0.72699071 0.9960796
#> 

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(
  fun = "lambda", arg = NA, lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.07084415 0.0000000 1.0000000 12.0000000 7.0000000
#> [2,] 0.09612690 0.2559460 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.2512618 0.4883274  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1046667  0.4692335 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.4664189 0.9890275
#> 
#> $A_list[[2]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.44453093 0.0000000 4.00000000 7.00000000 4.0000000
#> [2,] 0.01909926 0.2685161 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2993730 0.54188412 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.06983849 0.06694397 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.61958093 0.9355448
#> 
#> $A_list[[3]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.12790363 0.0000000 3.00000000 6.00000000 8.0000000
#> [2,] 0.01389691 0.2609287 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1555462 0.03941862 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.68500358 0.04804216 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.79593892 0.8767293
#> 
#> $A_list[[4]]
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.04276296 0.00000000 2.00000000 4.0000000 10.0000000
#> [2,] 0.07946151 0.06532181 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.47324939 0.55207664 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.03951404 0.1915398  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6524161  0.9795139
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]        [,4]      [,5]
#> [1,] 0.06532034 0.00000000 3.0000000 10.00000000 9.0000000
#> [2,] 0.24241975 0.33468433 0.0000000  0.00000000 0.0000000
#> [3,] 0.00000000 0.07295794 0.1602567  0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4480085  0.06224252 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.80364038 0.9177544
#> 
#> $A_list[[6]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.06330506 0.00000000 2.00000000 12.0000000 9.0000000
#> [2,] 0.42572577 0.48436024 0.00000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.02498285 0.01613295  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.58245676  0.5638603 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.3447892 0.9682582
#> 
#> $A_list[[7]]
#>            [,1]      [,2]      [,3]        [,4]      [,5]
#> [1,] 0.28130694 0.0000000 6.0000000 14.00000000 7.0000000
#> [2,] 0.01329849 0.3520649 0.0000000  0.00000000 0.0000000
#> [3,] 0.00000000 0.1370380 0.3821434  0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3391993  0.87903941 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.06741485 0.9562793
#> 
#> $A_list[[8]]
#>            [,1]      [,2]       [,3]      [,4]     [,5]
#> [1,] 0.50344882 0.0000000 7.00000000 6.0000000 8.000000
#> [2,] 0.01124001 0.4107161 0.00000000 0.0000000 0.000000
#> [3,] 0.00000000 0.2149330 0.73725625 0.0000000 0.000000
#> [4,] 0.00000000 0.0000000 0.05559021 0.6880850 0.000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.1665211 0.983266
#> 
#> $A_list[[9]]
#>            [,1]        [,2]      [,3]      [,4]       [,5]
#> [1,] 0.08456248 0.000000000 1.0000000 7.0000000 10.0000000
#> [2,] 0.31321867 0.008890238 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.417345595 0.5990419 0.0000000  0.0000000
#> [4,] 0.00000000 0.000000000 0.0291265 0.1879184  0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.4792748  0.8127577
#> 
#> $A_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09645311 0.0000000 5.0000000 6.0000000 7.0000000
#> [2,] 0.11067916 0.3205999 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2093930 0.6589105 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1950389 0.5597619 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3525394 0.9327064
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.07084415 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.09612690 0.2559460 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2512618 0.4883274 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1046667 0.4692335 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4664189 0.9890275
#> 
#> $U_list[[2]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.44453093 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.01909926 0.2685161 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.2993730 0.54188412 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.06983849 0.06694397 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.61958093 0.9355448
#> 
#> $U_list[[3]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.12790363 0.0000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.01389691 0.2609287 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1555462 0.03941862 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.68500358 0.04804216 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.79593892 0.8767293
#> 
#> $U_list[[4]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.04276296 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.07946151 0.06532181 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.47324939 0.55207664 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.03951404 0.1915398 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.6524161 0.9795139
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.06532034 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.24241975 0.33468433 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.07295794 0.1602567 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.4480085 0.06224252 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.80364038 0.9177544
#> 
#> $U_list[[6]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.06330506 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.42572577 0.48436024 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02498285 0.01613295 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.58245676 0.5638603 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.3447892 0.9682582
#> 
#> $U_list[[7]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.28130694 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.01329849 0.3520649 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1370380 0.3821434 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3391993 0.87903941 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.06741485 0.9562793
#> 
#> $U_list[[8]]
#>            [,1]      [,2]       [,3]      [,4]     [,5]
#> [1,] 0.50344882 0.0000000 0.00000000 0.0000000 0.000000
#> [2,] 0.01124001 0.4107161 0.00000000 0.0000000 0.000000
#> [3,] 0.00000000 0.2149330 0.73725625 0.0000000 0.000000
#> [4,] 0.00000000 0.0000000 0.05559021 0.6880850 0.000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.1665211 0.983266
#> 
#> $U_list[[9]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08456248 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.31321867 0.008890238 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.417345595 0.5990419 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.0291265 0.1879184 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.4792748 0.8127577
#> 
#> $U_list[[10]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09645311 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.11067916 0.3205999 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2093930 0.6589105 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1950389 0.5597619 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3525394 0.9327064
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1   12    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    7    4
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    6    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    4   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3   10    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2   12    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   14    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    6    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    7   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    6    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(
  fun = "eigs", arg = "lambda", lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>           [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,] 0.2161507 0.0000000 8.00000000 14.0000000 14.0000000
#> [2,] 0.1593418 0.2290627 0.00000000  0.0000000  0.0000000
#> [3,] 0.0000000 0.1547979 0.07291676  0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.36543165  0.2741112  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000  0.3679992  0.7232164
#> 
#> $A_list[[2]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.08615988 0.0000000 4.0000000 10.0000000 10.0000000
#> [2,] 0.11732682 0.1261017 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.1364704 0.5323726  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1707467  0.6829537  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.2591440  0.9806175
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]     [,5]
#> [1,] 0.39286606 0.0000000 2.0000000 6.00000000 9.000000
#> [2,] 0.05552539 0.3613480 0.0000000 0.00000000 0.000000
#> [3,] 0.00000000 0.4079863 0.5300627 0.00000000 0.000000
#> [4,] 0.00000000 0.0000000 0.2600125 0.76283509 0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.09758296 0.968145
#> 
#> $A_list[[4]]
#>           [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.0921426 0.00000000 1.00000000 7.00000000 16.0000000
#> [2,] 0.6762118 0.75336221 0.00000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.01632691 0.73178291 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.07970195 0.04821521  0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.78785713  0.8385334
#> 
#> $A_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.63491103 0.0000000 2.0000000 6.0000000 13.0000000
#> [2,] 0.03633678 0.6817623 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1008318 0.4661711 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3539158 0.4287638  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4504884  0.9389108
#> 
#> $A_list[[6]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.16396944 0.00000000 4.00000000 14.0000000 2.0000000
#> [2,] 0.09527945 0.01927343 0.00000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.53638234 0.57551679  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.05236781  0.1277600 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000  0.7490546 0.9826449
#> 
#> $A_list[[7]]
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.40039099 0.00000000 5.00000000 8.00000000 15.0000000
#> [2,] 0.00856624 0.34281328 0.00000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.09883649 0.02033964 0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.64901009 0.69647925  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.04139881  0.9174565
#> 
#> $A_list[[8]]
#>            [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.05333045 0.00000000 5.0000000 13.0000000 13.0000000
#> [2,] 0.26961423 0.51347338 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.02914928 0.6468313  0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.2580993  0.1649817  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.7839429  0.9808786
#> 
#> $A_list[[9]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.0106649 0.0000000 7.00000000 8.0000000 5.0000000
#> [2,] 0.3366602 0.2825883 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1057821 0.39028707 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.07953102 0.6682562 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2634544 0.9734692
#> 
#> $A_list[[10]]
#>            [,1]        [,2]      [,3]      [,4]       [,5]
#> [1,] 0.02667478 0.000000000 2.0000000 0.0000000 10.0000000
#> [2,] 0.07229794 0.689608719 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.009504626 0.3168677 0.0000000  0.0000000
#> [4,] 0.00000000 0.000000000 0.5811043 0.2420261  0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.7370713  0.9860218
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2161507 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1593418 0.2290627 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1547979 0.07291676 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.36543165 0.2741112 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.3679992 0.7232164
#> 
#> $U_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08615988 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.11732682 0.1261017 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1364704 0.5323726 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1707467 0.6829537 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2591440 0.9806175
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]     [,5]
#> [1,] 0.39286606 0.0000000 0.0000000 0.00000000 0.000000
#> [2,] 0.05552539 0.3613480 0.0000000 0.00000000 0.000000
#> [3,] 0.00000000 0.4079863 0.5300627 0.00000000 0.000000
#> [4,] 0.00000000 0.0000000 0.2600125 0.76283509 0.000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.09758296 0.968145
#> 
#> $U_list[[4]]
#>           [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.0921426 0.00000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.6762118 0.75336221 0.00000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.01632691 0.73178291 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.07970195 0.04821521 0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.78785713 0.8385334
#> 
#> $U_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.63491103 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03633678 0.6817623 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1008318 0.4661711 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3539158 0.4287638 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4504884 0.9389108
#> 
#> $U_list[[6]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.16396944 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.09527945 0.01927343 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.53638234 0.57551679 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.05236781 0.1277600 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.7490546 0.9826449
#> 
#> $U_list[[7]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.40039099 0.00000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.00856624 0.34281328 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.09883649 0.02033964 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.64901009 0.69647925 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.04139881 0.9174565
#> 
#> $U_list[[8]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05333045 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.26961423 0.51347338 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.02914928 0.6468313 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.2580993 0.1649817 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.7839429 0.9808786
#> 
#> $U_list[[9]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.0106649 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.3366602 0.2825883 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1057821 0.39028707 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.07953102 0.6682562 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2634544 0.9734692
#> 
#> $U_list[[10]]
#>            [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.02667478 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.07229794 0.689608719 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.009504626 0.3168677 0.0000000 0.0000000
#> [4,] 0.00000000 0.000000000 0.5811043 0.2420261 0.0000000
#> [5,] 0.00000000 0.000000000 0.0000000 0.7370713 0.9860218
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    8   14   14
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   10   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6    9
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    7   16
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   14    2
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5    8   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    5   13   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7    8    5
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    0   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.20544532 0.0000000 4.0000000 11.0000000 10.0000000
#> [2,] 0.03888898 0.3033266 0.0000000  0.0000000  0.0000000
#> [3,] 0.00000000 0.3188418 0.3885146  0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.5206746  0.3972484  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.5166877  0.9359775
#> 
#> $A_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.01870106 0.0000000 1.0000000 7.0000000 15.0000000
#> [2,] 0.49239952 0.4152797 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1324803 0.3024794 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.6103273 0.2732049  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7086761  0.9868661
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.04937309 0.0000000 3.0000000 8.00000000 7.0000000
#> [2,] 0.14600203 0.1103222 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.3324651 0.7581856 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1376251 0.07409553 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.83798101 0.9448672
#> 
#> $A_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]    [,5]
#> [1,] 0.34907963 0.0000000 2.0000000 6.0000000 8.00000
#> [2,] 0.01208253 0.5793715 0.0000000 0.0000000 0.00000
#> [3,] 0.00000000 0.1099380 0.5730888 0.0000000 0.00000
#> [4,] 0.00000000 0.0000000 0.1382164 0.7480443 0.00000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2001111 0.96609
#> 
#> $A_list[[5]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.08109296 0.0000000 6.0000000 13.0000000 8.0000000
#> [2,] 0.28289966 0.1261773 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.6422624 0.3260046  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4499055  0.3052328 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000  0.4870999 0.9708633
#> 
#> $A_list[[6]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1513748 0.0000000 6.0000000 8.0000000 3.0000000
#> [2,] 0.3681896 0.4315032 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3110814 0.7243782 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1248461 0.1207747 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.7831515 0.9753903
#> 
#> $A_list[[7]]
#>           [,1]      [,2]       [,3]        [,4]      [,5]
#> [1,] 0.3170758 0.0000000 3.00000000 6.000000000 7.0000000
#> [2,] 0.1274305 0.2898559 0.00000000 0.000000000 0.0000000
#> [3,] 0.0000000 0.1960889 0.57749909 0.000000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.04579787 0.008307365 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.710512842 0.8831219
#> 
#> $A_list[[8]]
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.08512308 0.00000000 3.00000000 8.0000000 13.0000000
#> [2,] 0.29317845 0.05399848 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.46298192 0.02135385 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.62909065 0.5694629  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1039028  0.7356154
#> 
#> $A_list[[9]]
#>            [,1]      [,2]       [,3]       [,4]      [,5]
#> [1,] 0.16180819 0.0000000 4.00000000 12.0000000 3.0000000
#> [2,] 0.05269955 0.4144130 0.00000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.2775315 0.06758881  0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.66103063  0.3678164 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000  0.4996111 0.9763838
#> 
#> $A_list[[10]]
#>            [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.31575557 0.00000000 7.0000000 14.0000000 7.0000000
#> [2,] 0.04014586 0.01177942 0.0000000  0.0000000 0.0000000
#> [3,] 0.00000000 0.57528074 0.3461818  0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3081779  0.3968673 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000  0.4114237 0.9515409
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.20544532 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.03888898 0.3033266 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3188418 0.3885146 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.5206746 0.3972484 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.5166877 0.9359775
#> 
#> $U_list[[2]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01870106 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.49239952 0.4152797 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.1324803 0.3024794 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.6103273 0.2732049 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7086761 0.9868661
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.04937309 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.14600203 0.1103222 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.3324651 0.7581856 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1376251 0.07409553 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.83798101 0.9448672
#> 
#> $U_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]    [,5]
#> [1,] 0.34907963 0.0000000 0.0000000 0.0000000 0.00000
#> [2,] 0.01208253 0.5793715 0.0000000 0.0000000 0.00000
#> [3,] 0.00000000 0.1099380 0.5730888 0.0000000 0.00000
#> [4,] 0.00000000 0.0000000 0.1382164 0.7480443 0.00000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2001111 0.96609
#> 
#> $U_list[[5]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08109296 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.28289966 0.1261773 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.6422624 0.3260046 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.4499055 0.3052328 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4870999 0.9708633
#> 
#> $U_list[[6]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1513748 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.3681896 0.4315032 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3110814 0.7243782 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.1248461 0.1207747 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.7831515 0.9753903
#> 
#> $U_list[[7]]
#>           [,1]      [,2]       [,3]        [,4]      [,5]
#> [1,] 0.3170758 0.0000000 0.00000000 0.000000000 0.0000000
#> [2,] 0.1274305 0.2898559 0.00000000 0.000000000 0.0000000
#> [3,] 0.0000000 0.1960889 0.57749909 0.000000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.04579787 0.008307365 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.710512842 0.8831219
#> 
#> $U_list[[8]]
#>            [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.08512308 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.29317845 0.05399848 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.46298192 0.02135385 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.62909065 0.5694629 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.1039028 0.7356154
#> 
#> $U_list[[9]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.16180819 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.05269955 0.4144130 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2775315 0.06758881 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.66103063 0.3678164 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.4996111 0.9763838
#> 
#> $U_list[[10]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.31575557 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.04014586 0.01177942 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.57528074 0.3461818 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.3081779 0.3968673 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.4114237 0.9515409
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   11   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    1    7   15
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    8    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    2    6    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6   13    8
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    6    8    3
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    6    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    3    8   13
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4   12    3
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    7   14    7
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>