Generate lists of Lefkovitch matrix population models (MPMs) based on life history archetypes

This function generates a list of n MPMs according to the specified criteria. Criteria include the archetype, and the acceptable constraining criteria, which could include lambda, generation time or any other metric derived from an A matrix. The function attempts to find matrices that fulfil the criteria, discarding unacceptable matrices. By default, if it takes more than 1000 attempts to find a suitable matrix model, then an error is produced. However, the number of attempts can be altered with the attempts parameter.

Usage

generate_mpm_set(
  n = 10,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  as_compadre = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)

Arguments

n

The number of MPMs to generate. Default is 10.

n_stages

The number of stages for the MPMs. Default is 3.

archetype

The archetype of the MPMs. Default is 1.

fecundity

A vector of fecundities for the MPMs. Default is 1.5.

split

A logical indicating whether to split into submatrices. Default is TRUE.

by_type

A logical indicating whether the matrices should be returned in a list by type (A, U, F, C). If split is FALSE, then by_type must is coerced to be FALSE. Defaults to TRUE.

as_compadre

A logical indicating whether the matrices should be returned as a CompadreDB object. Default is TRUE. If FALSE, the function returns a list.

max_surv

The maximum acceptable survival value, calculated across all transitions from a stage. Defaults to 0.99. This is only used if split = TRUE.

constraint

An optional data frame with 4 columns named fun, arg, lower and upper. These columns specify (1) a function that outputs a metric derived from an A matrix and (2) an argument for the function (NA, if no argument supplied) (3) the lower acceptable bound for the metric and (4) upper acceptable bound for the metric. This could be used to specify

attempts

An integer indicating the number of attempts To be made when simulating matrix model. The default is 1000. If it takes more than 1000 attempts to make a matrix that satisfies the conditions set by the other arguments, then a warning is produced.

Value

A list of MPMs that meet the specified criteria.

See also

Author

Owen Jones jones@biology.sdu.dk

Examples

set.seed(42) # set seed for repeatability

# Basic operation, without splitting matrices and with no constraints
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, split = FALSE, by_type = FALSE, as_compadre = FALSE
)
#> [[1]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.04028499 0.00000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.71907852 0.73066610 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.06797628 0.5129277 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.2895418 0.8589953  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1331578  0.9997924
#> 
#> [[2]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.04993404 0.00000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.35213116 0.38662704 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.04413311 0.2022055 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.4116914 0.4473679  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.4300040  0.9673925
#> 
#> [[3]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.3908713 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.1593942 0.3008004 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.4611680 0.5792468 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.2741919 0.2238051  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.6835265  0.9941976
#> 
#> [[4]]
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.01338708 0.00000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.46380038 0.47987890 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.03799584 0.64886531 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.01000717 0.3753701  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.3401858  0.7684915
#> 
#> [[5]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.67478475 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.04733024 0.4311501 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.4961576 0.3634766 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.5725528 0.7279782  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2540920  0.9967819
#> 
#> [[6]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.3819523 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.2431420 0.3941680 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.4929144 0.2382274 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.7382744 0.7182629  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.2756301  0.9947253
#> 
#> [[7]]
#>            [,1]      [,2]       [,3]       [,4]       [,5]
#> [1,] 0.16074427 0.0000000 4.00000000 8.00000000 10.0000000
#> [2,] 0.03843446 0.1920300 0.00000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.5599534 0.83063495 0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.01691095 0.06976802  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.92957038  0.9997121
#> 
#> [[8]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.11547275 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.09164534 0.3982819 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1783419 0.4565705 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3108012 0.7438372  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2286126  0.9994464
#> 
#> [[9]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.40726303 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.02472534 0.5507259 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.1762418 0.4303426 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3952927 0.1772480  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.7574622  0.9818421
#> 
#> [[10]]
#>             [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.195267196 0.00000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.009539193 0.01108828 0.0000000 0.0000000  0.0000000
#> [3,] 0.000000000 0.40598266 0.1292108 0.0000000  0.0000000
#> [4,] 0.000000000 0.00000000 0.3215483 0.2109751  0.0000000
#> [5,] 0.000000000 0.00000000 0.0000000 0.4432477  0.8537529
#> 

# Constrain outputs to A matrices with lambda between 0.9 and 1.1
library(popbio)
constrain_df <- data.frame(
  fun = "lambda", arg = NA, lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df, as_compadre = FALSE
)
#> $A_list
#> $A_list[[1]]
#>           [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2025021 0.00000000 4.00000000 8.0000000 10.000000
#> [2,] 0.1808410 0.51560726 0.00000000 0.0000000  0.000000
#> [3,] 0.0000000 0.05938074 0.69893955 0.0000000  0.000000
#> [4,] 0.0000000 0.00000000 0.04270201 0.1153077  0.000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.6563415  0.905483
#> 
#> $A_list[[2]]
#>           [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] 0.3492059 0.00000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.2623854 0.59588086 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.02204176 0.03327729 0.0000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.64881290 0.1459793  0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.6109135  0.9411863
#> 
#> $A_list[[3]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.0100986 0.0000000 4.00000000 8.0000000 10.000000
#> [2,] 0.2816060 0.3421699 0.00000000 0.0000000  0.000000
#> [3,] 0.0000000 0.2403473 0.60176000 0.0000000  0.000000
#> [4,] 0.0000000 0.0000000 0.01202352 0.5052573  0.000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2482741  0.975591
#> 
#> $A_list[[4]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.05277117 0.00000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.14635732 0.14395871 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.09478692 0.2766104 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.6115977 0.1849629  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.7926018  0.9830451
#> 
#> $A_list[[5]]
#>           [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2695917 0.00000000 4.0000000 8.0000000 10.000000
#> [2,] 0.2267870 0.45112437 0.0000000 0.0000000  0.000000
#> [3,] 0.0000000 0.05989263 0.3118949 0.0000000  0.000000
#> [4,] 0.0000000 0.00000000 0.3091554 0.2289075  0.000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.6927089  0.956369
#> 
#> $A_list[[6]]
#>           [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.2788292 0.0000000 4.0000000 8.00000000 10.0000000
#> [2,] 0.0816387 0.2433186 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.3089479 0.1000315 0.00000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.4670093 0.73745701  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.04978961  0.8887938
#> 
#> $A_list[[7]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2772425 0.0000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.1276093 0.4284850 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1104893 0.05017521 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.60407149 0.1454031  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.7206787  0.8721605
#> 
#> $A_list[[8]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.00794897 0.0000000 4.0000000 8.00000000 10.0000000
#> [2,] 0.10589532 0.2360636 0.0000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.1030530 0.2807865 0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1883105 0.62282892  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.08604455  0.9653569
#> 
#> $A_list[[9]]
#>           [,1]     [,2]       [,3]      [,4]       [,5]
#> [1,] 0.1495233 0.000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.1741336 0.426369 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.136707 0.62072487 0.0000000  0.0000000
#> [4,] 0.0000000 0.000000 0.06783148 0.1980795  0.0000000
#> [5,] 0.0000000 0.000000 0.00000000 0.5218206  0.7994516
#> 
#> $A_list[[10]]
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.04831774 0.00000000 4.00000000 8.00000000 10.0000000
#> [2,] 0.06775510 0.03037236 0.00000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.37177964 0.09800894 0.00000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.33187578 0.74345426  0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.01656937  0.7731065
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]       [,2]       [,3]      [,4]     [,5]
#> [1,] 0.2025021 0.00000000 0.00000000 0.0000000 0.000000
#> [2,] 0.1808410 0.51560726 0.00000000 0.0000000 0.000000
#> [3,] 0.0000000 0.05938074 0.69893955 0.0000000 0.000000
#> [4,] 0.0000000 0.00000000 0.04270201 0.1153077 0.000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.6563415 0.905483
#> 
#> $U_list[[2]]
#>           [,1]       [,2]       [,3]      [,4]      [,5]
#> [1,] 0.3492059 0.00000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.2623854 0.59588086 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.02204176 0.03327729 0.0000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.64881290 0.1459793 0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.6109135 0.9411863
#> 
#> $U_list[[3]]
#>           [,1]      [,2]       [,3]      [,4]     [,5]
#> [1,] 0.0100986 0.0000000 0.00000000 0.0000000 0.000000
#> [2,] 0.2816060 0.3421699 0.00000000 0.0000000 0.000000
#> [3,] 0.0000000 0.2403473 0.60176000 0.0000000 0.000000
#> [4,] 0.0000000 0.0000000 0.01202352 0.5052573 0.000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.2482741 0.975591
#> 
#> $U_list[[4]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.05277117 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.14635732 0.14395871 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.09478692 0.2766104 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.6115977 0.1849629 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.7926018 0.9830451
#> 
#> $U_list[[5]]
#>           [,1]       [,2]      [,3]      [,4]     [,5]
#> [1,] 0.2695917 0.00000000 0.0000000 0.0000000 0.000000
#> [2,] 0.2267870 0.45112437 0.0000000 0.0000000 0.000000
#> [3,] 0.0000000 0.05989263 0.3118949 0.0000000 0.000000
#> [4,] 0.0000000 0.00000000 0.3091554 0.2289075 0.000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.6927089 0.956369
#> 
#> $U_list[[6]]
#>           [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.2788292 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.0816387 0.2433186 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.3089479 0.1000315 0.00000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.4670093 0.73745701 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.04978961 0.8887938
#> 
#> $U_list[[7]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2772425 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1276093 0.4284850 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1104893 0.05017521 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.60407149 0.1454031 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.7206787 0.8721605
#> 
#> $U_list[[8]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.00794897 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.10589532 0.2360636 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1030530 0.2807865 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1883105 0.62282892 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.08604455 0.9653569
#> 
#> $U_list[[9]]
#>           [,1]     [,2]       [,3]      [,4]      [,5]
#> [1,] 0.1495233 0.000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1741336 0.426369 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.136707 0.62072487 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000 0.06783148 0.1980795 0.0000000
#> [5,] 0.0000000 0.000000 0.00000000 0.5218206 0.7994516
#> 
#> $U_list[[10]]
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.04831774 0.00000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.06775510 0.03037236 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.37177964 0.09800894 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.33187578 0.74345426 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.01656937 0.7731065
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# As above, but using popdemo::eigs function instead of popbio::lambda
# to illustrate use of argument
library(popdemo)
#> Welcome to popdemo! This is version 1.3-0
#> Use ?popdemo for an intro, or browseVignettes('popdemo') for vignettes
#> Citation for popdemo is here: doi.org/10.1111/j.2041-210X.2012.00222.x
#> Development and legacy versions are here: github.com/iainmstott/popdemo
constrain_df <- data.frame(
  fun = "eigs", arg = "lambda", lower = 0.9, upper =
    1.1
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df, as_compadre = FALSE
)
#> $A_list
#> $A_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.33328577 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.06377826 0.2498128 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3337334 0.5963528 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.1693849 0.5795012  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2212670  0.8103362
#> 
#> $A_list[[2]]
#>           [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.0239377 0.00000000 4.00000000 8.00000000 10.0000000
#> [2,] 0.0366482 0.04343731 0.00000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.68395612 0.68478775 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.08214221 0.06041167  0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.82870382  0.9361283
#> 
#> $A_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]       [,5]
#> [1,] 0.03354061 0.0000000 4.0000000 8.00000000 10.0000000
#> [2,] 0.24257199 0.1737578 0.0000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.1323195 0.2086060 0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3520131 0.61162941  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.06355464  0.7634507
#> 
#> $A_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.18298502 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.05516901 0.0637293 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3666609 0.3365942 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.3601747 0.2711895  0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4848073  0.9864156
#> 
#> $A_list[[5]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.00572345 0.0000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.25867680 0.1849481 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.2763296 0.47875737 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.04194771 0.2885805  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5292559  0.8627223
#> 
#> $A_list[[6]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1689392 0.00000000 4.0000000 8.00000000 10.0000000
#> [2,] 0.2063822 0.30596227 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.07818665 0.1925666 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.3427810 0.01826475  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.63827242  0.9756126
#> 
#> $A_list[[7]]
#>            [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.00652138 0.0000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.00545548 0.2302399 0.00000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.3754500 0.65618224 0.0000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.01014227 0.6273468  0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3247017  0.9748711
#> 
#> $A_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.1884203 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.1412555 0.5157395 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.1011372 0.3102842 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3591664 0.6594515  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.1373212  0.8799535
#> 
#> $A_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.09476255 0.00000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.07831552 0.54792290 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.09611724 0.5001257 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.1557204 0.5926258  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1258167  0.8132592
#> 
#> $A_list[[10]]
#>            [,1]      [,2]     [,3]       [,4]       [,5]
#> [1,] 0.05344835 0.0000000 4.000000 8.00000000 10.0000000
#> [2,] 0.19325338 0.1600322 0.000000 0.00000000  0.0000000
#> [3,] 0.00000000 0.1377219 0.301253 0.00000000  0.0000000
#> [4,] 0.00000000 0.0000000 0.336472 0.65045350  0.0000000
#> [5,] 0.00000000 0.0000000 0.000000 0.08209705  0.8151836
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.33328577 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.06377826 0.2498128 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3337334 0.5963528 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.1693849 0.5795012 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.2212670 0.8103362
#> 
#> $U_list[[2]]
#>           [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.0239377 0.00000000 0.00000000 0.00000000 0.0000000
#> [2,] 0.0366482 0.04343731 0.00000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.68395612 0.68478775 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.08214221 0.06041167 0.0000000
#> [5,] 0.0000000 0.00000000 0.00000000 0.82870382 0.9361283
#> 
#> $U_list[[3]]
#>            [,1]      [,2]      [,3]       [,4]      [,5]
#> [1,] 0.03354061 0.0000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.24257199 0.1737578 0.0000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1323195 0.2086060 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3520131 0.61162941 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.06355464 0.7634507
#> 
#> $U_list[[4]]
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.18298502 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.05516901 0.0637293 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3666609 0.3365942 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.3601747 0.2711895 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.4848073 0.9864156
#> 
#> $U_list[[5]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.00572345 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.25867680 0.1849481 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2763296 0.47875737 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.04194771 0.2885805 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.5292559 0.8627223
#> 
#> $U_list[[6]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1689392 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.2063822 0.30596227 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.07818665 0.1925666 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.3427810 0.01826475 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.63827242 0.9756126
#> 
#> $U_list[[7]]
#>            [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.00652138 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.00545548 0.2302399 0.00000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.3754500 0.65618224 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.01014227 0.6273468 0.0000000
#> [5,] 0.00000000 0.0000000 0.00000000 0.3247017 0.9748711
#> 
#> $U_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.1884203 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1412555 0.5157395 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.1011372 0.3102842 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3591664 0.6594515 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.1373212 0.8799535
#> 
#> $U_list[[9]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.09476255 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.07831552 0.54792290 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.09611724 0.5001257 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1557204 0.5926258 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.1258167 0.8132592
#> 
#> $U_list[[10]]
#>            [,1]      [,2]     [,3]       [,4]      [,5]
#> [1,] 0.05344835 0.0000000 0.000000 0.00000000 0.0000000
#> [2,] 0.19325338 0.1600322 0.000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.1377219 0.301253 0.00000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.336472 0.65045350 0.0000000
#> [5,] 0.00000000 0.0000000 0.000000 0.08209705 0.8151836
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> 

# Multiple constraints
# Constrain outputs to A matrices with lambda between 0.9 and 1.1, generation
# time between 3 and 5 and damping ratio between 1 and 7.
library(popbio)
constrain_df <- data.frame(
  fun = c("lambda", "generation.time", "damping.ratio"),
  arg = c(NA, NA, NA),
  lower = c(0.9, 3.0, 1.0),
  upper = c(1.1, 5.0, 7.0)
)
generate_mpm_set(
  n = 10, n_stages = 5, fecundity = c(0, 0, 4, 8, 10),
  archetype = 4, constraint = constrain_df, as_compadre = FALSE
)
#> $A_list
#> $A_list[[1]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2215161 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.1628780 0.2398348 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.4176027 0.5451896 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.2235974 0.2776922  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.5255625  0.9682391
#> 
#> $A_list[[2]]
#>           [,1]        [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2633244 0.000000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.2108113 0.532344404 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.004494054 0.1274034 0.0000000  0.0000000
#> [4,] 0.0000000 0.000000000 0.5361963 0.1105499  0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.5715804  0.8004599
#> 
#> $A_list[[3]]
#>             [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.036019608 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.007165058 0.1339785 0.0000000 0.0000000  0.0000000
#> [3,] 0.000000000 0.4912836 0.7179049 0.0000000  0.0000000
#> [4,] 0.000000000 0.0000000 0.1357687 0.4106779  0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.5146034  0.9286131
#> 
#> $A_list[[4]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2242099 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.4212880 0.2347628 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.4932071 0.5667867 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.3834741 0.4861952  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4657886  0.9670306
#> 
#> $A_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 0.01632312 0.00000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.16488654 0.21741849 0.0000000 0.0000000  0.0000000
#> [3,] 0.00000000 0.04855655 0.3199929 0.0000000  0.0000000
#> [4,] 0.00000000 0.00000000 0.1293828 0.3595308  0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.4448042  0.8932768
#> 
#> $A_list[[6]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.4144249 0.0000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.1257273 0.3536676 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.2470648 0.73727607 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.06073812 0.1518907  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.6807092  0.8370672
#> 
#> $A_list[[7]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1172182 0.00000000 4.0000000 8.00000000 10.0000000
#> [2,] 0.2959306 0.58901722 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.04150199 0.1807843 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.6355638 0.07568457  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.90359482  0.9831803
#> 
#> $A_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 0.2851186 0.0000000 4.0000000 8.0000000 10.0000000
#> [2,] 0.4914640 0.5399783 0.0000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.3387130 0.2868371 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.6329236 0.6625216  0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.3037399  0.9721035
#> 
#> $A_list[[9]]
#>           [,1]      [,2]       [,3]      [,4]       [,5]
#> [1,] 0.2791730 0.0000000 4.00000000 8.0000000 10.0000000
#> [2,] 0.2425753 0.3766383 0.00000000 0.0000000  0.0000000
#> [3,] 0.0000000 0.2745754 0.83467374 0.0000000  0.0000000
#> [4,] 0.0000000 0.0000000 0.01164569 0.4305712  0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.4928974  0.9476323
#> 
#> $A_list[[10]]
#>           [,1]       [,2]      [,3]       [,4]       [,5]
#> [1,] 0.1053867 0.00000000 4.0000000 8.00000000 10.0000000
#> [2,] 0.4436386 0.53789037 0.0000000 0.00000000  0.0000000
#> [3,] 0.0000000 0.02305909 0.1275618 0.00000000  0.0000000
#> [4,] 0.0000000 0.00000000 0.7625399 0.08617495  0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.81587337  0.9278515
#> 
#> 
#> $U_list
#> $U_list[[1]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2215161 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.1628780 0.2398348 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.4176027 0.5451896 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.2235974 0.2776922 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.5255625 0.9682391
#> 
#> $U_list[[2]]
#>           [,1]        [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2633244 0.000000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.2108113 0.532344404 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.004494054 0.1274034 0.0000000 0.0000000
#> [4,] 0.0000000 0.000000000 0.5361963 0.1105499 0.0000000
#> [5,] 0.0000000 0.000000000 0.0000000 0.5715804 0.8004599
#> 
#> $U_list[[3]]
#>             [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.036019608 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.007165058 0.1339785 0.0000000 0.0000000 0.0000000
#> [3,] 0.000000000 0.4912836 0.7179049 0.0000000 0.0000000
#> [4,] 0.000000000 0.0000000 0.1357687 0.4106779 0.0000000
#> [5,] 0.000000000 0.0000000 0.0000000 0.5146034 0.9286131
#> 
#> $U_list[[4]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2242099 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.4212880 0.2347628 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.4932071 0.5667867 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.3834741 0.4861952 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.4657886 0.9670306
#> 
#> $U_list[[5]]
#>            [,1]       [,2]      [,3]      [,4]      [,5]
#> [1,] 0.01632312 0.00000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.16488654 0.21741849 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.04855655 0.3199929 0.0000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.1293828 0.3595308 0.0000000
#> [5,] 0.00000000 0.00000000 0.0000000 0.4448042 0.8932768
#> 
#> $U_list[[6]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.4144249 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.1257273 0.3536676 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2470648 0.73727607 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.06073812 0.1518907 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.6807092 0.8370672
#> 
#> $U_list[[7]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1172182 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.2959306 0.58901722 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.04150199 0.1807843 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.6355638 0.07568457 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.90359482 0.9831803
#> 
#> $U_list[[8]]
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2851186 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.4914640 0.5399783 0.0000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.3387130 0.2868371 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.6329236 0.6625216 0.0000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.3037399 0.9721035
#> 
#> $U_list[[9]]
#>           [,1]      [,2]       [,3]      [,4]      [,5]
#> [1,] 0.2791730 0.0000000 0.00000000 0.0000000 0.0000000
#> [2,] 0.2425753 0.3766383 0.00000000 0.0000000 0.0000000
#> [3,] 0.0000000 0.2745754 0.83467374 0.0000000 0.0000000
#> [4,] 0.0000000 0.0000000 0.01164569 0.4305712 0.0000000
#> [5,] 0.0000000 0.0000000 0.00000000 0.4928974 0.9476323
#> 
#> $U_list[[10]]
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 0.1053867 0.00000000 0.0000000 0.00000000 0.0000000
#> [2,] 0.4436386 0.53789037 0.0000000 0.00000000 0.0000000
#> [3,] 0.0000000 0.02305909 0.1275618 0.00000000 0.0000000
#> [4,] 0.0000000 0.00000000 0.7625399 0.08617495 0.0000000
#> [5,] 0.0000000 0.00000000 0.0000000 0.81587337 0.9278515
#> 
#> 
#> $F_list
#> $F_list[[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> $F_list[[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    4    8   10
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#>