Generate random matrix population models (MPMs) based on life history archetypes

Generates random matrix population models (MPMs) with element values based on defined life history archetypes. Survival and transition/growth probabilities from any particular stage are restricted to be less than or equal to 1 by drawing from a Dirichlet distribution. The user can specify archetypes (from Takada et al. 2018) to restrict the MPMs in other ways:

Archetype 1: all elements are positive, although they may be very small. Therefore, transition from/to any stage is possible. This model describes a life history where individuals can progress and retrogress rapidly.

Archetype 2: has the same form as archetype 1 (transition from/to any stage is possible), but the survival probability (column sums of the survival matrix) increases monotonously as the individuals advance to later stages. This model, as the one in the first archetype, also allows for rapid progression and retrogression, but is more realistic in that stage-specific survival probability increases with stage advancement.

Archetype 3: positive non-zero elements for survival are only allowed on the diagonal and lower sub-diagonal of the matrix This model represents the life cycle of a species where retrogression is not allowed, and progression can only happen to the immediately larger/more developed stage (slow progression, e.g., trees).

Archetype 4: This archetype has the same general form as archetype 3, but with the further assumption that stage-specific survival increases as individuals increase in size/developmental stage. In this respect it is similar to archetype 2.

Usage

random_mpm(n_stages, fecundity, archetype = 1, split = FALSE)

Arguments

n_stages

An integer defining the number of stages for the MPM.

fecundity

Mean fecundity. This value is the lambda value for a Poisson distribution from which a value for fecundity is drawn. An integer of length 1 or a vector of integers of length equal to the number of stages. If there is no reproduction in a particular age class, use a value of 0.

archetype

Indication of which life history archetype should be used, based on Takada et al. 2018. An integer between 1 and 4.

split

TRUE/FALSE, indicating whether the matrix produced should be split into a survival matrix and a fertility matrix. Yeah true, then the output becomes a list with a matrix in each element. Otherwise, the output is a single matrix.

Value

Returns a random matrix population model with characteristics determined by the archetype selected and fecundity vector. If split = TRUE, the matrix is split into separate fertility and a growth/survival matrices, returned as a list.

Details

In all 4 of these Archetypes, fecundity is placed as a single element on the top right of the matrix, if it is a single value. If it is a vector of length n_stages then the fertility vector spans the entire top row of the matrix.

References

Takada, T., Kawai, Y., & Salguero-Gómez, R. (2018). A cautionary note on elasticity analyses in a ternary plot using randomly generated population matrices. Population Ecology, 60(1), 37–47.

Author

Owen Jones jones@biology.sdu.dk

Examples

random_mpm(n_stages = 2, fecundity = 20, archetype = 1, split = FALSE)
#>            [,1]      [,2]
#> [1,] 0.08103769 8.5466243
#> [2,] 0.31521824 0.1083336
random_mpm(n_stages = 2, fecundity = 20, archetype = 2, split = TRUE)
#> $matU
#>           [,1]      [,2]
#> [1,] 0.6608433 0.1006448
#> [2,] 0.2155896 0.8929039
#> 
#> $matF
#>      [,1] [,2]
#> [1,]    0   20
#> [2,]    0    0
#> 
random_mpm(n_stages = 3, fecundity = 20, archetype = 3, split = FALSE)
#>           [,1]       [,2]       [,3]
#> [1,] 0.1003168 0.00000000 19.0000000
#> [2,] 0.5671734 0.01970339  0.0000000
#> [3,] 0.0000000 0.78240286  0.8257511
random_mpm(n_stages = 4, fecundity = 20, archetype = 4, split = TRUE)
#> $matU
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.3246496 0.0000000 0.0000000 0.0000000
#> [2,] 0.2330349 0.1910844 0.0000000 0.0000000
#> [3,] 0.0000000 0.4819316 0.2780015 0.0000000
#> [4,] 0.0000000 0.0000000 0.4993682 0.8938325
#> 
#> $matF
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0   16
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 
random_mpm(
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4,
  split = TRUE
)
#> $matU
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.08101457 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,] 0.20298852 0.1259955 0.0000000 0.0000000 0.0000000
#> [3,] 0.00000000 0.2047407 0.3187331 0.0000000 0.0000000
#> [4,] 0.00000000 0.0000000 0.5457502 0.6067151 0.0000000
#> [5,] 0.00000000 0.0000000 0.0000000 0.3403059 0.9526292
#> 
#> $matF
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    5   11
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#>